refactor(cli): découper le point d'entrée par public (0.1.69)
cli.py était passé de point d'entrée à fourre-tout : 3 289 lignes et 34
commandes, alors que la logique métier vit ailleurs. Pas un défaut de conception
mais de trajectoire, chaque commande nouvelle y ajoutant ses options, sa
validation et son orchestration. catalog, infra et status, livrés le même jour,
le rendaient difficile à repousser encore.

Quatorze modules désormais : contexte, parcours, indices et progression pour
l'apprenant ; infrastructure, destruction et etat ; catalogues ; auteur,
instructeur et diagnostic ; plus _socle, _commun et _barres pour ce qu'ils
partagent.

Rien ne bouge pour qui est dehors. dsoxlab.cli:main reste le point d'entrée
déclaré, from dsoxlab.cli import app fonctionne toujours, et dsoxlab --help
liste les mêmes commandes DANS LE MÊME ORDRE. Ce dernier point est le piège du
découpage : Typer affiche les commandes dans leur ordre d'enregistrement, donc
répartir le code les réordonnait sans qu'aucun test ne le dise, alors que
--help est la première chose que lit un utilisateur. L'ordre est maintenant une
décision explicite, comparée avant et après par diff.

Le découpage a montré au passage que trois helpers de barres de progression,
utilisés aussi par provision et run, s'étaient retrouvés collés après destroy :
c'est ce qui donnait à ce bloc ses 435 lignes. Ils ont leur module.

Sept tests changent d'ADRESSE, pas de contenu : ils patchaient dsoxlab.cli.X, or
patcher l'attribut d'un paquet ne change pas ce qu'un module utilise. Aucun
n'est réécrit, aucune assertion ne bouge. Le garde-fou i18n, lui, suit le
découpage comme l'issue le demandait : il nomme désormais des modules de publics
différents, pour qu'un découpage futur qui en oublierait un se voie.

Écart connu, assumé : _commun.py reste à 701 lignes, au-dessus des 400 exigées.
Sa scission a été tentée puis ANNULÉE, parce qu'elle cassait le paquet entier et
que les réparations s'enchaînaient sans converger. Le paquet a été reconstruit
d'un seul tenant depuis l'original, par un script déterministe. Cette scission
mérite sa propre modification, avec les tests verts à chaque étape.

Vérifié : 705 tests unitaires, 18 e2e, ruff, mypy strict sur 76 fichiers, ordre
de --help identique au diff près, et la CLI jouée sur deux dépôts fournisseurs.

Closes #119
