--- src/dsoxlab/cli.py
@@ -1,3289 +1,0 @@
-"""Point d'entrée CLI — dsoxlab.
-
-Usage:
-    dsoxlab use linux/l1
-    dsoxlab list-labs
-    dsoxlab show <id>
-    dsoxlab run <id>
-    dsoxlab check <id>
-    dsoxlab reset <id>
-    dsoxlab clean <id>
-    dsoxlab validate-structure
-    dsoxlab doctor
-    dsoxlab quit
-
-Convention de ce module : un ``except`` qui a déjà rendu la cause en une phrase
-traduite (``error(...)``) sort par ``raise typer.Exit(n) from None``. Le ``from
-None`` n'est pas un raccourci, c'est l'affirmation que la cause a été dite à
-l'utilisateur, et qu'un chaînage d'exceptions n'ajouterait qu'une trace Python
-au-dessus d'un message déjà écrit pour lui. Partout ailleurs, on chaîne.
-"""
-
-from __future__ import annotations
-
-import atexit
-import logging
-import os
-import shutil
-import subprocess
-import sys
-import webbrowser
-from collections.abc import Callable
-from pathlib import Path
-from typing import Annotated, Any, NoReturn
-
-import typer
-from typer.core import TyperGroup, TyperOption
-
-from . import __version__
-from .config import (
-    clear_context,
-    get_lab_home,
-    read_context,
-    set_active_lab,
-    set_active_provider,
-    set_course_pos,
-    write_context,
-)
-from .i18n import _, get_lang, set_lang
-from .infra import libvirt
-from .infra.inventory import InfraNotProvisioned
-from .interrupt import (
-    EVENT_INTERRUPT,
-    EXIT_INTERRUPTED,
-    Interrupted,
-    Stage,
-    interruptible,
-)
-from .locking import EXIT_LOCKED, RepoLock, RepoLocked
-from .logging_setup import configurer as configurer_journal
-from .models import (
-    ContractError,
-    CourseManifest,
-    LabDefinition,
-    ProviderUnresolved,
-    RepoMetadata,
-    UnsupportedSchemaVersion,
-)
-from .models.hint import HintFile
-from .reporting import (
-    console,
-    error,
-    info,
-    machine,
-    paged,
-    print_check_result,
-    print_course_end,
-    print_course_section,
-    print_course_toc,
-    print_doctor,
-    print_fullhelp,
-    print_hint,
-    print_lab_challenge,
-    print_lab_course,
-    print_lab_detail,
-    print_lab_welcome,
-    print_labs_table,
-    print_progress_table,
-    print_scores_table,
-    print_structure_reports,
-    success,
-    update_console,
-    warn,
-)
-from .runtimes.base import EventCallback
-from .services import (
-    CheckResult,
-    check_lab,
-    clean_lab,
-    collect_checks,
-    evaluate_lab,
-    find_lab,
-    get_all_labs,
-    guide_url,
-    host_diagnosis,
-    lab_status,
-    next_pending_lab,
-    open_lab_session,
-    reset_lab,
-    run_lab,
-    validate_all_metadata,
-    validate_all_structure,
-)
-from .sessions.store import (
-    get_best_scores,
-    get_results,
-    hints_cost_total,
-    next_hint_index,
-    record_hint,
-    reset_hints,
-)
-from .utils.shell import CommandError
-
-logger = logging.getLogger(__name__)
-
-
-class _I18nGroup(TyperGroup):
-    """TyperGroup avec l'option --help traduite."""
-
-    # ``ctx`` est annoté ``Any`` volontairement : typer a fait évoluer le
-    # type du Context de get_help_option (click public en 0.25, copie
-    # vendorée typer._click en 0.26+). Annoter ``Any`` garde la surcharge
-    # valide (LSP) sans coupler le code à un module privé qui n'existe pas
-    # dans toutes les versions. Le retour ``TyperOption`` est public et
-    # stable.
-    def get_help_option(self, ctx: Any) -> TyperOption | None:
-        opt = super().get_help_option(ctx)
-        if opt is not None:
-            opt.help = _("opt_help")
-        return opt
-
-    def invoke(self, ctx: Any) -> Any:
-        """Traduit un Ctrl-C resté sans propriétaire en interruption nommée.
-
-        C'est le **seul** endroit qui puisse le faire. Typer attrape lui-même
-        ``KeyboardInterrupt`` tout en bas de son ``_main()`` et le change en
-        ``Exit(130)`` : le code de retour était déjà juste, mais l'interruption
-        y perdait toute identité, et un filet posé autour de ``app()`` ne voit
-        jamais rien passer. ``invoke`` s'exécute à l'intérieur de ce ``_main()``,
-        donc en amont : c'est le dernier moment où l'on peut encore dire à
-        l'apprenant ce qui vient d'être interrompu, au lieu de lui rendre son
-        invite sans un mot.
-
-        On convertit ici, et **jamais** à la source par un handler de signal :
-        ``KeyboardInterrupt`` descend de ``BaseException``, ce qui est
-        exactement ce qui lui permet de traverser les ``except Exception`` que
-        ce code pose autour des callbacks d'affichage. Une exception ordinaire
-        levée depuis un handler s'y ferait avaler en silence, et le flux
-        Terraform continuerait comme si de rien n'était.
-        """
-        try:
-            return super().invoke(ctx)
-        except KeyboardInterrupt:
-            raise Interrupted(Stage.UNKNOWN) from None
-        except FileNotFoundError as exc:
-            # Un binaire absent remontait jusqu'à l'interpréteur, et une trace
-            # Python dit « l'outil est cassé » alors qu'il manque le plus
-            # souvent un paquet que l'apprenant peut poser lui-même. Toutes les
-            # autres erreurs attendues sont déjà converties en message + code ;
-            # celle-ci ne l'était pas, faute d'un endroit qui la voie passer.
-            manquant = exc.filename or str(exc)
-            # Un nom sans séparateur a été cherché dans le PATH : c'est la
-            # définition d'un exécutable introuvable, et 127 est le code que le
-            # shell rend dans ce cas. Un chemin, lui, désigne un fichier.
-            est_executable = "/" not in str(manquant)
-            error(_(
-                "err_executable_introuvable" if est_executable
-                else "err_fichier_introuvable",
-                nom=manquant,
-            ))
-            raise typer.Exit(127 if est_executable else 2) from None
-
-
-app = typer.Typer(
-    name="dsoxlab",
-    help=_("app_help"),
-    no_args_is_help=True,
-    rich_markup_mode="rich",
-    add_completion=True,
-    cls=_I18nGroup,
-)
-
-# ── Sous-application 'instructor' (commandes formateur) ───────────────────────
-
-instructor_app = typer.Typer(
-    name="instructor",
-    help=_("cmd_instructor_help"),
-    no_args_is_help=True,
-    rich_markup_mode="rich",
-    cls=_I18nGroup,
-)
-app.add_typer(instructor_app, name="instructor")
-
-# ── Sous-application 'completion' ─────────────────────────────────────────────
-
-completion_app = typer.Typer(
-    name="completion",
-    help=_("cmd_completion_help"),
-    no_args_is_help=True,
-    rich_markup_mode="rich",
-    cls=_I18nGroup,
-)
-app.add_typer(completion_app, name="completion")
-
-# ── Sous-application 'catalog' ────────────────────────────────────────────────
-
-catalog_app = typer.Typer(
-    name="catalog",
-    help=_("cmd_catalog_help"),
-    no_args_is_help=True,
-    rich_markup_mode="rich",
-    cls=_I18nGroup,
-)
-app.add_typer(catalog_app, name="catalog")
-
-# ── Sous-application 'infra' ──────────────────────────────────────────────────
-
-infra_app = typer.Typer(
-    name="infra",
-    help=_("cmd_infra_help"),
-    no_args_is_help=True,
-    rich_markup_mode="rich",
-    cls=_I18nGroup,
-)
-app.add_typer(infra_app, name="infra")
-
-# ── Option globale lab-home ───────────────────────────────────────────────────
-
-LabHomeOption = Annotated[
-    Path | None,
-    typer.Option(
-        "--lab-home",
-        envvar="LAB_HOME",
-        help=_("opt_lab_home"),
-        show_default=False,
-    ),
-]
-
-
-#: Les affichages longs (cours, mission de challenge) passent par le pager
-#: dès qu'ils dépassent l'écran. L'option permet de retrouver le déversement
-#: brut, par exemple pour copier tout un README d'un seul tenant.
-NoPagerOption = Annotated[
-    bool,
-    typer.Option("--no-pager", help=_("opt_no_pager")),
-]
-
-
-def _root(lab_home: Path | None) -> Path:
-    return lab_home.resolve() if lab_home else get_lab_home()
-
-
-def _contrat_trop_recent(exc: UnsupportedSchemaVersion) -> NoReturn:
-    """Rend l'erreur de version du ``meta.yml`` et sort.
-
-    Le ``meta.yml`` décrit tout le catalogue : ne pas savoir le lire ne laisse
-    rien de fiable derrière. On ne dégrade donc pas, on nomme la cause et la
-    réparation, qui n'est pas dans le dépôt de labs mais dans la version de
-    l'outil.
-    """
-    error(_(
-        "schema_version_meta_too_new",
-        path=exc.source, found=exc.found, supported=exc.supported,
-    ))
-    raise typer.Exit(1)
-
-
-def _phrase_contrat(exc: ContractError) -> str:
-    """La phrase d'un champ du contrat que le moteur ne sait pas lire.
-
-    Le modèle porte les faits (``source``, ``field``) et la **clé** ; la phrase
-    se compose ici, dans la langue de l'apprenant. Le chemin du fichier encadre
-    le message plutôt que d'entrer dans chaque traduction : il est le même pour
-    toutes, et le répéter dans les tables les ferait diverger.
-
-    Un seul endroit la compose, parce que deux appelants la disent :
-    :func:`_contrat_illisible`, qui sort en ``typer.Exit``, et le filet de
-    :func:`main`, qui sort en ``SystemExit``.
-    """
-    return f"{exc.source}: {_(exc.key, **exc.params)}"
-
-
-def _contrat_illisible(exc: ContractError) -> NoReturn:
-    """Rend l'erreur de contrat et sort, pour les commandes qui lisent le catalogue."""
-    error(_phrase_contrat(exc))
-    raise typer.Exit(1)
-
-
-def _read_repo(root: Path) -> RepoMetadata | None:
-    """Wrapper de ``read_repo_metadata`` qui formate proprement les
-    erreurs de résolution de provider (ValueError) en message CLI +
-    typer.Exit(1), au lieu d'un traceback Python brut.
-    """
-    from .discovery.repo import read_repo_metadata
-
-    try:
-        return read_repo_metadata(root)
-    # AVANT ValueError, dont elles héritent toutes deux : `str(exc)` rendrait
-    # un message technique et non traduit là où la cause mérite une phrase.
-    except UnsupportedSchemaVersion as exc:
-        _contrat_trop_recent(exc)
-    # Le meta.yml est le seul fichier du contrat dont les erreurs de lecture
-    # s'affichent : le modèle porte la clé et les faits, la phrase se compose
-    # dans `_contrat_illisible`.
-    except ContractError as exc:
-        _contrat_illisible(exc)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-
-def _catalogue(root: Path, lang: str, *, quiet: bool = False) -> list[LabDefinition]:
-    """Le catalogue, et un mot sur ce qu'il a fallu en écarter.
-
-    Un ``lab.yaml`` au contrat trop récent est laissé de côté — c'est la seule
-    chose honnête à en faire — mais jamais en silence : sans cet
-    avertissement, la seule trace serait un lab absent, le symptôme le plus
-    coûteux à diagnostiquer de tout le contrat.
-
-    ``quiet`` sert au mode ``--json``, où la sortie standard ne doit porter
-    qu'un document et rien d'autre.
-    """
-    from .discovery.scanner import scan_catalog
-
-    try:
-        scan = scan_catalog(root, lang=lang)
-    except UnsupportedSchemaVersion as exc:
-        _contrat_trop_recent(exc)
-    # Le scanner relit le meta.yml pour l'ordre des sections : un champ mal
-    # typé y remontait donc en traceback, sur `list-labs` comme sur `progress`,
-    # alors que `_read_repo` disait déjà la phrase sur les autres commandes.
-    except ContractError as exc:
-        _contrat_illisible(exc)
-    if not quiet:
-        for ecarte in scan.unsupported:
-            warn(_(
-                "schema_version_lab_skipped",
-                path=ecarte.source, found=ecarte.found, supported=ecarte.supported,
-            ))
-    return scan.labs
-
-
-def _lab(root: Path, lab_id: str, lang: str) -> LabDefinition:
-    """Un lab par son identifiant, en nommant d'abord ce qui a été écarté.
-
-    Sans cela, ``dsoxlab show <id>`` d'un lab au contrat trop récent répondrait
-    « lab introuvable », et laisserait croire à une faute de frappe plutôt qu'à
-    un outil à mettre à jour.
-    """
-    return find_lab(_catalogue(root, lang), lab_id)
-
-
-def _require_provider(repo_meta: RepoMetadata) -> str:
-    """Retourne le provider actif, ou sort proprement s'il n'est pas résolu.
-
-    À n'appeler que dans les commandes d'infrastructure (provision,
-    destroy, status…). Les commandes pédagogiques (list-labs, show,
-    check, scores…) n'ont pas besoin de provider et ne doivent pas
-    échouer quand le dépôt en déclare plusieurs.
-    """
-    try:
-        return repo_meta.infra.require_provider()
-    except ProviderUnresolved as exc:
-        if not exc.candidates:
-            error(_("provider_none_declared"))
-        else:
-            error(_("provider_required",
-                    candidates=", ".join(exc.candidates),
-                    first=exc.candidates[0]))
-        raise typer.Exit(1) from None
-
-
-def _verrou(root: Path, commande: str) -> RepoLock:
-    """Prend le verrou d'écriture du dépôt, ou sort en nommant qui le tient.
-
-    À n'appeler que dans les commandes qui **modifient** l'état. Les commandes
-    de lecture (``list-labs``, ``show``, ``scores``, ``progress``, ``next``,
-    ``status``, ``doctor``, ``course``, ``challenge``, ``hint``, ``guide``,
-    ``validate-structure``, ``support``, ``fullhelp``) ne passent pas par ici :
-    consulter son catalogue pendant qu'un ``provision`` tourne dans un autre
-    terminal est un usage normal, pas un conflit.
-
-    Le verrou rendu est **déjà pris**. L'appelant choisit sa portée :
-
-    - ``ctx.call_on_close(_verrou(...).release)`` pour toute la commande ;
-    - ``with _verrou(...):`` pour la seule phase qui écrit. C'est ce que fait
-      ``run``, qui doit rendre le verrou **avant** d'ouvrir la session
-      interactive, sinon le ``dsoxlab check`` que l'apprenant tape dans ce
-      sous-shell serait refusé par sa propre session.
-    """
-    verrou = RepoLock(root, commande)
-    try:
-        verrou.acquire()
-    except RepoLocked as exc:
-        detenteur = exc.holder
-        if detenteur is None:
-            error(_("lock_busy_anonymous", path=exc.path))
-        else:
-            error(_(
-                "lock_busy",
-                command=detenteur.command, pid=detenteur.pid,
-                age=detenteur.age_label,
-            ))
-        info(_("lock_busy_hint"))
-        raise typer.Exit(EXIT_LOCKED) from None
-    return verrou
-
-
-def _interrompu(exc: Interrupted, reprise: str) -> NoReturn:
-    """Rend une interruption : ce qui reste en place, puis comment reprendre.
-
-    Trois affirmations, dans cet ordre, et le code de sortie **130** qui les
-    confirme (``128 + SIGINT``) : ce qui a été interrompu, ce que ça laisse
-    derrière, et la commande qui reprend. Le code, typer le rendait déjà ; ce
-    qui manquait, c'était la phrase, et sur le chemin Ansible le code lui-même,
-    qui annonçait un échec (2) là où l'apprenant avait appuyé sur Ctrl-C.
-    """
-    warn(_(exc.message_key))
-    if exc.hard:
-        warn(_("interrupted_hard"))
-    info(_("interrupt_resume", cmd=reprise))
-    raise typer.Exit(EXIT_INTERRUPTED)
-
-
-def _services_repo_id(root: Path) -> str:
-    """Slug du dépôt, pour namespacer les conteneurs de services.
-
-    Le ``meta.yml`` fournit ``repo.id`` ; à défaut, le nom du répertoire racine.
-    """
-    from .discovery.repo import read_repo_metadata
-    try:
-        meta = read_repo_metadata(root)
-    except ValueError:
-        meta = None
-    return meta.id if meta else root.name
-
-
-def _ensure_services(lab: LabDefinition, root: Path) -> None:
-    """Démarre les services conteneurisés déclarés par le lab, avant run/check.
-
-    Sans service déclaré, ne fait rien. Si Docker est injoignable ou qu'un
-    service ne démarre pas, sort proprement : un lab qui déclare un service ne
-    peut pas fonctionner sans lui.
-    """
-    if not lab.runtime.services:
-        return
-    from .runtimes import services as svc
-    if not svc.docker_available():
-        error(_("services_docker_absent"))
-        raise typer.Exit(2)
-    repo_id = _services_repo_id(root)
-    for service in lab.runtime.services:
-        info(_("service_starting", name=service.name, image=service.image))
-        # Le démarrage attend les sondes du service, jusqu'à `ready_timeout`
-        # secondes : c'est le point d'attente le plus long d'un lab shell, donc
-        # celui où le Ctrl-C tombe.
-        with interruptible(Stage.SERVICES):
-            try:
-                svc.start(service, repo_id)
-            except svc.ServiceError as exc:
-                error(_("service_failed", name=service.name, detail=str(exc)))
-                raise typer.Exit(2) from None
-        success(_("service_ready", name=service.name))
-
-
-def _stop_services(lab: LabDefinition, root: Path) -> None:
-    """Arrête les services conteneurisés du lab, à destroy/clean. Best-effort."""
-    if not lab.runtime.services:
-        return
-    from .runtimes import services as svc
-    if not svc.docker_available():
-        return
-    repo_id = _services_repo_id(root)
-    for service in lab.runtime.services:
-        if svc.stop(service, repo_id):
-            info(_("service_stopped", name=service.name))
-
-
-def _lang(root: Path) -> str:
-    """Langue effective : contexte > DSOXLAB_LANG > LANG système > en."""
-    ctx = read_context(root)
-    return get_lang(ctx_lang=ctx.lang)
-
-
-def _complete_lab_id(incomplete: str) -> list[tuple[str, str]]:
-    """Propose les lab IDs du dépôt courant, filtrés par ce qui est déjà saisi.
-
-    Contrat **typer**, pas celui de click. Deux différences, et elles ne sont
-    pas cosmétiques :
-
-    - typer construit lui-même les paramètres passés ici, d'après les
-      annotations de cette signature (``incomplete: str`` reçoit le préfixe
-      saisi). Il n'attend donc pas les trois positionnels
-      ``(ctx, param, incomplete)`` de click, et ce qu'on ne déclare pas n'est
-      pas injecté ;
-    - le retour est une liste de couples ``(valeur, aide)``. C'est typer qui en
-      fait des ``CompletionItem`` : lui rendre l'objet click directement casse
-      son adaptateur, qui exige une chaîne ou un couple.
-
-    Le second élément du couple est l'aide affichée à côté de la proposition,
-    c'est-à-dire le titre du lab. Elle survit à la migration, et zsh l'affiche.
-    """
-    try:
-        root = get_lab_home()
-        labs = get_all_labs(root)
-        return [
-            (lab.id, lab.title)
-            for lab in labs
-            if lab.id.startswith(incomplete)
-        ]
-    # Aveugle, et délibérément : une complétion s'exécute à chaque Tab, dans un
-    # répertoire quelconque. Un meta.yml absent, un lab.yaml qui lève, un
-    # catalogue à moitié écrit — rien de tout cela ne doit faire cracher une
-    # trace Python dans le shell de l'apprenant. Pas de proposition, et c'est tout.
-    except Exception:  # noqa: BLE001
-        return []
-
-
-def _version_callback(value: bool) -> None:
-    """Affiche la version puis quitte (option ``--version`` eager)."""
-    if value:
-        console.print(f"dsoxlab {__version__}")
-        raise typer.Exit()
-
-
-@app.callback(invoke_without_command=True)
-def _bootstrap(
-    ctx: typer.Context,
-    version: Annotated[
-        bool,
-        typer.Option(
-            "--version",
-            help=_("opt_version_help"),
-            callback=_version_callback,
-            is_eager=True,
-        ),
-    ] = False,
-    verbose: Annotated[
-        int,
-        typer.Option("--verbose", "-v", count=True, help=_("opt_verbose")),
-    ] = 0,
-    debug: Annotated[
-        bool,
-        typer.Option("--debug", help=_("opt_debug")),
-    ] = False,
-) -> None:
-    """Initialise la langue UI et la journalisation avant toute commande."""
-    # Avant le retour anticipé : `dsoxlab -v` sans sous-commande doit tout de
-    # même écrire son journal, et c'est aussi ce qui garantit qu'une commande
-    # qui échoue très tôt laisse une trace.
-    configurer_journal(verbose, debug=debug)
-
-    if ctx.invoked_subcommand is None:
-        return
-    try:
-        root = get_lab_home()
-        lang = _lang(root)
-        set_lang(lang)
-    # Aveugle et silencieux, volontairement : choisir la langue est un préalable
-    # à TOUTES les commandes. Sans contexte de lab, on continue en langue par
-    # défaut ; échouer ici empêcherait jusqu'à `dsoxlab --help`.
-    except Exception:  # noqa: S110, BLE001
-        pass  # silencieux si LAB_HOME introuvable
-
-    # L'avis de mise à jour est posé ici, mais affiché à la toute fin par
-    # atexit : c'est le seul moyen qu'il soit le dernier message, y compris
-    # quand la commande sort en erreur ou lève typer.Exit.
-    atexit.register(_notify_update_available)
-
-
-def _notify_update_available() -> None:
-    """Affiche l'avis de mise à jour, en dernier, sur stderr.
-
-    Sur stderr et pas stdout : une commande en `--json` doit rendre un
-    document lisible par un programme, quoi qu'il arrive. Et seulement si
-    stderr est un terminal, pour ne pas polluer les journaux d'une CI ni la
-    sortie capturée par un script.
-    """
-    if not sys.stderr.isatty():
-        return
-    try:
-        from .services.update_check import available_update
-
-        latest = available_update(__version__)
-        if latest is None:
-            return
-        update_console.print(
-            _("update_available", latest=latest, current=__version__)
-        )
-    # Aveugle, et c'est le but : un avis de mise à jour ne casse jamais la
-    # commande que l'utilisateur a lancée, quelle que soit la panne réseau,
-    # de parsing ou d'affichage rencontrée.
-    except Exception:  # noqa: BLE001
-        return
-
-
-# ── install ─────────────────────────────────────────────────────────────────
-
-#: Nom du programme, tel que la CLI est installée et invoquée.
-_PROG_NAME = "dsoxlab"
-
-#: Variable d'environnement par laquelle le shell demande une complétion.
-#: Elle est DÉRIVÉE du nom du programme, comme le fait Click lui-même, et non
-#: recopiée : la valeur codée en dur était « _DSOXL_COMPLETE », que la CLI
-#: n'écoute pas. Le script généré interrogeait donc dsoxlab avec une variable
-#: ignorée, la CLI répondait par sa page d'aide, et le shell tentait de
-#: l'évaluer à chaque tabulation.
-_COMPLETE_VAR = f"_{_PROG_NAME.replace('-', '_').upper()}_COMPLETE"
-
-
-#: Ce que dsoxlab ajoute au script que typer génère pour zsh, et pourquoi.
-#:
-#: Le commentaire part DANS le fichier installé, à dessein : sans lui, la ligne
-#: ressemble à une scorie qu'un lecteur pressé retirerait, et le défaut
-#: reviendrait sans que personne ne comprenne pourquoi.
-_ZSH_PREMIER_TAB = """
-# ── ajouté par dsoxlab, et pas par typer ──────────────────────────────────────
-# zsh autoload ce fichier au PREMIER Tab, et attend qu'il produise les
-# propositions de cette invocation-là. Le script amont se contente de définir la
-# fonction puis de l'enregistrer pour la suite : la première tabulation ne rend
-# donc rien, et la seconde fonctionne. Un Tab muet se lit comme « la complétion
-# ne marche pas », et personne ne rappuie pour vérifier.
-# Ne pas retirer cette ligne sans rejouer le cas dans un zsh réel.
-_dsoxlab_completion "$@"
-"""
-
-
-def _script_completion(shell: str) -> str:
-    """Le script de complétion pour ``shell``, corrigé pour zsh (#134)."""
-    from typer.completion import get_completion_script
-
-    script = get_completion_script(
-        prog_name=_PROG_NAME, complete_var=_COMPLETE_VAR, shell=shell
-    )
-    # bash source son script au démarrage du shell, fish le charge par fichier
-    # de complétion : ni l'un ni l'autre ne passe par l'autoload qui pose
-    # problème. La divergence ne vaut donc que pour zsh.
-    return script + _ZSH_PREMIER_TAB if shell == "zsh" else script
-
-
-def _installer_completion() -> None:
-    """Pose le script de complétion du shell courant, et raccorde son rc."""
-    shell_name = Path(os.environ.get("SHELL", "bash")).name
-
-    if shell_name == "zsh":
-        zfunc_dir = Path.home() / ".zfunc"
-        zfunc_dir.mkdir(exist_ok=True)
-        # Le nom du fichier compte : zsh autoload la fonction `_dsoxlab` pour
-        # compléter `dsoxlab`, et cherche donc un fichier de ce nom exact.
-        comp_file = zfunc_dir / f"_{_PROG_NAME}"
-        comp_file.write_text(_script_completion("zsh"))
-        success(_("install_completion", path=str(comp_file)))
-
-        zshrc = Path.home() / ".zshrc"
-        zshrc_content = zshrc.read_text() if zshrc.exists() else ""
-        additions = []
-        if "fpath=(~/.zfunc $fpath)" not in zshrc_content:
-            additions.append("fpath=(~/.zfunc $fpath)")
-        if "autoload -Uz compinit" not in zshrc_content:
-            additions.append("autoload -Uz compinit && compinit")
-        if additions:
-            with zshrc.open("a") as f:
-                f.write("\n# dsoxlab completion\n" + "\n".join(additions) + "\n")
-        info(_("install_rc", path=str(zshrc)))
-
-    elif shell_name == "bash":
-        bash_comp_dir = Path.home() / ".bash_completion.d"
-        bash_comp_dir.mkdir(exist_ok=True)
-        comp_file = bash_comp_dir / "dsoxlab"
-        comp_file.write_text(_script_completion("bash"))
-        success(_("install_completion", path=str(comp_file)))
-
-        bashrc = Path.home() / ".bashrc"
-        source_line = f". {comp_file}"
-        bashrc_content = bashrc.read_text() if bashrc.exists() else ""
-        if source_line not in bashrc_content:
-            with bashrc.open("a") as f:
-                f.write(f"\n# dsoxlab completion\n{source_line}\n")
-        info(_("install_rc", path=str(bashrc)))
-
-    else:
-        info(_("install_completion_unsupported", shell=shell_name))
-    info(_("install_reload"))
-
-
-@completion_app.command("install", help=_("cmd_completion_install_help"))
-def completion_install() -> None:
-    _installer_completion()
-
-
-@completion_app.command("show", help=_("cmd_completion_show_help"))
-def completion_show(
-    shell: Annotated[str | None, typer.Option("--shell", help=_("opt_completion_shell"))] = None,
-) -> None:
-    """Imprime le script, sans rien écrire : à rediriger où l'on veut."""
-    nom = shell or Path(os.environ.get("SHELL", "bash")).name
-    if nom not in ("zsh", "bash", "fish"):
-        error(_("install_completion_unsupported", shell=nom))
-        raise typer.Exit(2)
-    # `print` et non `info` : c'est une sortie destinée à être redirigée, elle
-    # ne doit porter ni couleur ni encadrement.
-    print(_script_completion(nom))
-
-
-@app.command("install", help=_("cmd_install_help"))
-def install() -> None:
-    """Déprécié depuis 0.1.62, retiré en 0.3.0 : voir `completion install`.
-
-    Le nom promettait d'installer l'outil, déjà installé. La commande posait en
-    plus un wrapper dans ``~/.local/bin``, exactement où ``uv tool install`` et
-    ``pipx`` posent le leur : le remplacer ne faisait que défaire ce que leur
-    prochaine mise à jour remettrait. Il n'est donc plus écrit du tout.
-    """
-    warn(_("install_deprecie"))
-    _installer_completion()
-
-
-# ── use ──────────────────────────────────────────────────────────────────────
-
-@app.command("use", help=_("cmd_use_help"))
-def use(
-    ctx: typer.Context,
-    context: Annotated[str | None, typer.Argument(help=_("cmd_use_arg"))] = None,
-    lab_home: LabHomeOption = None,
-    lang: Annotated[str | None, typer.Option("--lang", help=_("opt_lang"))] = None,
-    target: Annotated[str | None, typer.Option("--target", "-t",
-        help=_("opt_target"))] = None,
-    provider: Annotated[str | None, typer.Option("--provider", "-p",
-        help=_("opt_use_provider"))] = None,
-    reset: Annotated[bool, typer.Option("--reset", "-r", help=_("opt_use_reset"))] = False,
-) -> None:
-    root = _root(lab_home)
-    # `.dsoxlab-context.json` est réécrit EN ENTIER à chaque changement : deux
-    # `use` concurrents, et le premier est perdu sans laisser de trace.
-    ctx.call_on_close(_verrou(root, "use").release)
-    if reset:
-        clear_context(root)
-        success(_("context_cleared"))
-        return
-    # Si seule l'option --target est donnée (pas de section/level),
-    # on met juste à jour la target sans toucher au contexte.
-    if context is None and target is None and lang is None and provider is None:
-        clear_context(root)
-        success(_("context_cleared"))
-        return
-
-    # Lecture brute du meta.yml (pas _resolve_provider) : on doit pouvoir
-    # valider même quand la résolution standard laisse le provider non
-    # résolu (plusieurs candidats sans choix actif).
-    declared_providers: list[str] = []
-    declared_sections: list[str] = []
-    import yaml as _yaml
-
-    from .discovery.repo import find_meta_yml
-
-    meta_path = find_meta_yml(root) or (root / "meta.yml")
-    if meta_path.is_file():
-        try:
-            raw = _yaml.safe_load(meta_path.read_text(encoding="utf-8")) or {}
-        except _yaml.YAMLError as exc:
-            error(_("meta_read_failed", error=exc))
-            raise typer.Exit(1) from None
-        declared = (raw.get("infra") or {}).get("provider")
-        if isinstance(declared, list):
-            declared_providers = [str(p) for p in declared if p]
-        elif isinstance(declared, str) and declared:
-            declared_providers = [declared]
-        declared_sections = [
-            str(s.get("id")) for s in (raw.get("sections") or []) if s.get("id")
-        ]
-
-    # Garde-fou : `dsoxlab use incus` pose une SECTION nommée « incus »,
-    # pas un provider — et filtre alors le catalogue sur une section qui
-    # n'existe pas (« Aucun lab trouvé »). Piège d'autant plus vicieux que
-    # le nom ressemble à un provider. On refuse et on guide.
-    if (
-        context is not None
-        and context in declared_providers
-        and context not in declared_sections
-    ):
-        error(_("provider_not_a_section", name=context))
-        raise typer.Exit(1)
-
-    # Même piège, cas général : une section inconnue était acceptée sans un
-    # mot, puis « list-labs » répondait « Aucun lab trouvé ». L'apprenant
-    # croyait le catalogue vide alors qu'il venait de poser un filtre qui ne
-    # correspond à rien. On refuse, et on montre ce qui existe.
-    # Un meta.yml sans bloc « sections » ne déclare rien : on ne filtre pas.
-    if context is not None and declared_sections:
-        demandee = context.strip().split("/", 1)[0]
-        if demandee and demandee not in declared_sections:
-            error(_("section_unknown",
-                    name=demandee,
-                    sections=", ".join(declared_sections)))
-            raise typer.Exit(1)
-
-    # --provider <name> : valider contre les providers candidats du
-    # meta.yml avant de l'enregistrer dans le contexte session.
-    if provider is not None:
-        if declared_providers and provider not in declared_providers:
-            error(_("provider_unknown",
-                    name=provider,
-                    candidates=", ".join(declared_providers)))
-            raise typer.Exit(1)
-        set_active_provider(root, provider)
-        success(_("context_provider_set", provider=provider))
-
-    section: str | None = None
-    level: str | None = None
-    if context:
-        parts = context.strip().split("/", 1)
-        section = parts[0] or None
-        level = parts[1] if len(parts) > 1 else None
-    if context or lang or target:
-        write_context(root, section, level, lang=lang, active_target=target)
-    if section:
-        label = f"{section}/{level}" if level else section
-        success(_("context_set", label=label))
-        info(_("context_set_info"))
-    if lang:
-        info(_("context_lang_set", lang=lang[:2].lower()))
-    if target:
-        success(_("context_target_set", target=target))
-
-
-
-# ── list-labs ─────────────────────────────────────────────────────────────────
-
-@app.command("list-labs", help=_("cmd_list_labs_help"))
-def list_labs(
-    lab_home: LabHomeOption = None,
-    level: Annotated[str | None, typer.Option("--level", "-l", help=_("opt_level"))] = None,
-    section: Annotated[str | None, typer.Option("--section", "-s", help=_("opt_section"))] = None,
-    lab_type: Annotated[str | None, typer.Option("--type", "-t", help=_("opt_type"))] = None,
-    bloc: Annotated[int | None, typer.Option("--bloc", "-b", help=_("opt_bloc"))] = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    root = _root(lab_home)
-    ctx = read_context(root)
-
-    effective_section = section or ctx.section
-    effective_level = level or ctx.level
-
-    # En mode machine, aucun message d'ambiance : la sortie doit être un
-    # document JSON et rien d'autre.
-    if (ctx.section or ctx.level) and not as_json:
-        info(_("context_active", label=ctx.label()))
-
-    lang = _lang(root)
-    labs = _catalogue(root, lang, quiet=as_json)
-    if effective_section:
-        labs = [lab for lab in labs if lab.section == effective_section]
-    if effective_level:
-        labs = [lab for lab in labs if lab.level == effective_level]
-    if lab_type:
-        labs = [lab for lab in labs if lab.lab_type == lab_type]
-    if bloc is not None:
-        labs = [lab for lab in labs if lab.bloc == bloc]
-    lab_ids = [lab.id for lab in labs]
-    scores = get_best_scores(root, lab_ids)
-    if as_json:
-        machine.emit({
-            "labs": [machine.lab_dict(lab, scores.get(lab.id)) for lab in labs],
-            "count": len(labs),
-        })
-        return
-    print_labs_table(labs, scores)
-
-
-# ── show ──────────────────────────────────────────────────────────────────────
-
-@app.command("show", help=_("cmd_show_help"))
-def show(
-    lab_id: Annotated[str, typer.Argument(help=_("cmd_show_arg"), autocompletion=_complete_lab_id)],
-    lab_home: LabHomeOption = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    root = _root(lab_home)
-    lang = _lang(root)
-    try:
-        lab = _lab(root, lab_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    # Le drapeau et non une comparaison de chaînes : `status` est un jeton
-    # (`ready` / `stopped`) sauf dans ce cas-là, où c'est une phrase traduite.
-    # Le mode machine doit distinguer les deux sans lire du français.
-    indisponible = False
-    try:
-        status = lab_status(lab)
-    except RuntimeError:
-        status = _("runtime_unavailable")
-        indisponible = True
-
-    if as_json:
-        # `best_score` doit dire la vérité : le laisser à `null` sur un lab
-        # déjà noté se lirait « jamais tenté ».
-        scores = get_best_scores(root, [lab.id])
-        machine.emit({
-            "lab": machine.lab_dict(lab, scores.get(lab.id)),
-            "status": None if indisponible else status,
-        })
-        return
-
-    print_lab_detail(lab, status=status)
-
-
-# ── run ───────────────────────────────────────────────────────────────────────
-
-@app.command("run", help=_("cmd_run_help"))
-def run(
-    lab_id: Annotated[str, typer.Argument(help=_("cmd_run_arg"), autocompletion=_complete_lab_id)],
-    target: Annotated[str | None, typer.Option("--target", "-t",
-        help=_("opt_run_target"))] = None,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lang = _lang(root)
-    try:
-        lab = _lab(root, lab_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    info(_("lab_starting", lab_id=lab.id, runtime=lab.runtime.type.value))
-    # Le verrou ne couvre QUE la phase qui écrit : services, playbook de setup,
-    # contexte. Il est rendu avant la session interactive, sinon le `dsoxlab
-    # check` que l'apprenant tape dans ce sous-shell serait refusé par sa
-    # propre session, blocage que personne ne comprendrait.
-    with _verrou(root, "run"):
-        _ensure_services(lab, root)
-        try:
-            if lab.runtime.type.value in ("vm", "kvm", "incus"):
-                _run_ansible_with_progress(
-                    lab.path / "setup.yaml",
-                    lambda cb: run_lab(lab, target_name=target, on_event=cb, root=root),
-                )
-            else:
-                run_lab(lab, target_name=target, root=root)
-        except Interrupted as exc:
-            _interrompu(exc, f"dsoxlab run {lab.id}")
-        except InfraNotProvisioned:
-            # Avant le except RuntimeError : InfraNotProvisioned en hérite, et
-            # mérite la phrase actionnable plutôt que son message technique.
-            error(_("infra_not_provisioned"))
-            raise typer.Exit(2) from None
-        except RuntimeError as exc:
-            error(str(exc))
-            raise typer.Exit(2) from None
-
-        set_active_lab(root, lab.id)
-    # Dire où l'on atterrit vraiment. Le message historique annonçait
-    # « challenge/work/ » quel que soit le runtime : faux pour tout lab vm,
-    # qui ouvre une session SSH ou, désormais, un shell à la racine du dépôt.
-    if lab.runtime.type.value in ("vm", "kvm", "incus"):
-        if lab.runtime.session == "local":
-            success(_("lab_ready_local", lab_id=lab.id))
-        else:
-            resolved = lab.runtime.target(target)
-            success(_("lab_ready_target", lab_id=lab.id,
-                      host=resolved.host if resolved else "?"))
-    else:
-        success(_("lab_ready", lab_id=lab.id, workdir=lab.runtime.workdir))
-    # La session SSH s'ouvre sur un hôte dépourvu de dsoxlab : une fois dedans,
-    # l'apprenant ne peut plus afficher sa mission. On la lui met sous les yeux
-    # avant d'entrer, elle reste dans le défilement du terminal.
-    if lab.runtime.type.value in ("vm", "kvm", "incus") and lab.runtime.session != "local":
-        print_lab_challenge(lab, lang=lang)
-
-    print_lab_welcome(lab)
-
-    try:
-        open_lab_session(lab)   # bloquant : sous-shell interactif
-    except Interrupted as exc:
-        _interrompu(exc, f"dsoxlab run {lab.id}")
-
-    # Retour au shell parent : on garde active_lab posé pour que
-    # ``dsoxlab check`` et ``dsoxlab submit`` (sans argument) sachent
-    # quel lab valider. L'active_lab est libéré au submit (cf. cmd
-    # submit) ou écrasé par un prochain ``dsoxlab run <autre_lab>``.
-    # En session locale, on n'a jamais quitté son répertoire : annoncer un
-    # « retour » n'aurait aucun sens. Ce qui compte alors, c'est que le travail
-    # reste là et que check puisse être relancé.
-    if lab.runtime.session == "local":
-        success(_("lab_session_ended_local", lab_id=lab.id))
-    else:
-        success(_("lab_session_ended", lab_id=lab.id))
-
-
-
-
-# ── course ───────────────────────────────────────────────────────────────────
-
-@app.command("course", help=_("cmd_course_help"))
-def course(
-    lab_id: Annotated[str | None, typer.Argument(help=_("cmd_course_arg"), autocompletion=_complete_lab_id)] = None,
-    section: Annotated[str | None, typer.Option("--section", "-s", help=_("cmd_course_opt_section"))] = None,
-    next_section: Annotated[bool, typer.Option("--next", "-n", help=_("cmd_course_opt_next"))] = False,
-    prev_section: Annotated[bool, typer.Option("--prev", "-p", help=_("cmd_course_opt_prev"))] = False,
-    no_pager: NoPagerOption = False,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lang = _lang(root)
-    lab = _resolve_lab(root, lab_id, lang)
-
-
-    manifest = CourseManifest.load(lab.path, lang=lang)
-
-    if manifest is None:
-        # Pas de course.yaml : fallback sur scenario.md + README.md, soit
-        # plusieurs centaines de lignes d'un bloc. C'est le cas qui rend le
-        # pager indispensable.
-        with paged(enabled=not no_pager):
-            print_lab_course(lab, lang=lang)
-        return
-
-    total = len(manifest.sections)
-    ctx = read_context(root)
-    current_pos = ctx.course_pos  # 0 = non démarré
-
-    # ── Résolution de la position cible ──────────────────────────────────────
-    if next_section:
-        target_pos = (current_pos or 0) + 1
-    elif prev_section:
-        target_pos = max(1, (current_pos or 1) - 1)
-    elif section is not None:
-        found = manifest.resolve_section(section)
-        if found is None:
-            error(_("course_section_not_found", name=section, id=lab.id))
-            raise typer.Exit(1)
-        # Retrouver l'index 1-based de la section
-        target_pos = next(
-            (i + 1 for i, s in enumerate(manifest.sections) if s.id == found.id),
-            1,
-        )
-    else:
-        # Pas d'option : reprendre là où on en était, ou afficher section 1
-        if current_pos and 1 <= current_pos <= total:
-            target_pos = current_pos
-        elif current_pos and current_pos > total:
-            print_course_end(lab, manifest)
-            return
-        else:
-            # Jamais commencé → sommaire + section 1, paginés ensemble : les
-            # séparer ferait défiler le sommaire hors de l'écran.
-            with paged(enabled=not no_pager):
-                print_course_toc(lab, manifest)
-                _show_course_section(root, lab, manifest, 1)
-            return
-
-    # ── Fin de cours ─────────────────────────────────────────────────────────
-    if target_pos > total:
-        set_course_pos(root, total)
-        print_course_end(lab, manifest)
-        return
-
-    # ── Affichage de la section ───────────────────────────────────────────────
-    with paged(enabled=not no_pager):
-        _show_course_section(root, lab, manifest, target_pos)
-
-
-def _show_course_section(
-    root: Path, lab: LabDefinition, manifest: CourseManifest, pos: int
-) -> None:
-    """Mémorise la position atteinte puis affiche la section correspondante."""
-    total = len(manifest.sections)
-    set_course_pos(root, pos)
-    print_course_section(lab, manifest.sections[pos - 1], pos=pos, total=total)
-
-
-# ── challenge ─────────────────────────────────────────────────────────────────
-
-@app.command("challenge", help=_("cmd_challenge_help"))
-def challenge_cmd(
-    lab_id: Annotated[str | None, typer.Argument(help=_("cmd_challenge_arg"), autocompletion=_complete_lab_id)] = None,
-    no_pager: NoPagerOption = False,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lab = _resolve_lab(root, lab_id, _lang(root))
-    with paged(enabled=not no_pager):
-        print_lab_challenge(lab, lang=_lang(root))
-
-
-# ── hint ──────────────────────────────────────────────────────────────────────
-
-@app.command("guide", help=_("cmd_guide_help"))
-def guide(
-    lab_id: Annotated[str | None, typer.Argument(help=_("cmd_guide_arg"), autocompletion=_complete_lab_id)] = None,
-    print_only: Annotated[bool, typer.Option("--print", help=_("cmd_guide_opt_print"))] = False,
-    lab_home: LabHomeOption = None,
-) -> None:
-    """Ouvre le guide en ligne du lab dans le navigateur.
-
-    Le cours est publié sur le site du formateur, pas embarqué dans le dépôt :
-    on ouvre donc la vraie page plutôt que d'en rapatrier le contenu. Elle
-    s'affiche telle qu'elle est publiée, et la lecture compte comme une visite
-    réelle du site.
-    """
-    root = _root(lab_home)
-    lang = _lang(root)
-    lab = _resolve_lab(root, lab_id, lang)
-
-    url = guide_url(lab)
-    if url is None:
-        error(_("guide_no_url", lab_id=lab.id))
-        raise typer.Exit(1)
-
-    # soft_wrap : une URL coupée sur deux lignes n'est plus copiable ni
-    # exploitable dans un pipe. Elle doit sortir d'un seul tenant, même
-    # au-delà de la largeur du terminal.
-    if print_only:
-        console.print(url, soft_wrap=True)
-        return
-
-    info(_("guide_opening", lab_id=lab.id))
-    console.print(url, soft_wrap=True)
-    # L'URL reste affichée : sur une machine sans navigateur (session SSH,
-    # serveur), webbrowser rend False sans rien ouvrir, et l'apprenant doit
-    # pouvoir la copier.
-    if not webbrowser.open(url):
-        error(_("guide_no_browser"))
-
-
-@app.command("hint", help=_("cmd_hint_help"))
-def hint(
-    lab_id: Annotated[str | None, typer.Argument(help=_("cmd_hint_arg"), autocompletion=_complete_lab_id)] = None,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lang = _lang(root)
-    ctx = read_context(root)
-    effective_id = lab_id or ctx.active_lab
-    if not effective_id:
-        error(_("no_active_lab"))
-        raise typer.Exit(1)
-    try:
-        lab = _lab(root, effective_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    hint_file = HintFile.load(lab.path / "challenge")
-    if not hint_file.hints:
-        info(_("no_hints"))
-        return
-
-    idx = next_hint_index(root, effective_id)
-    if idx >= len(hint_file.hints):
-        info(_("all_hints_used", count=len(hint_file.hints), total=len(hint_file.hints)))
-        return
-
-    current = hint_file.hints[idx]
-    record_hint(root, effective_id, idx, current.cost)
-    total_cost = hints_cost_total(root, effective_id)
-    print_hint(
-        idx, len(hint_file.hints),
-        current.text(_lang(root)),
-        current.cost, total_cost,
-    )
-
-
-# ── check helpers ─────────────────────────────────────────────────────────────
-
-def _resolve_lab(
-    root: Path, lab_id: str | None, lang: str
-) -> LabDefinition:
-    """Résout l'ID effectif et retourne le LabDefinition, ou lève typer.Exit."""
-    ctx = read_context(root)
-    effective_id = lab_id or ctx.active_lab
-    if not effective_id:
-        error(_("no_active_lab"))
-        raise typer.Exit(1)
-    try:
-        return _lab(root, effective_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-
-def _run_check_with_progress(
-    lab: LabDefinition, target: str | None = None, *, quiet: bool = False,
-) -> CheckResult:
-    """Lance ``check_lab`` en streamant les verdicts pytest dans une
-    progress bar Rich.
-
-    ``target`` sélectionne la target du lab sur laquelle valider (labs
-    multi-distrib). None = la target ``default`` du lab.
-
-    Affiche un ✔/✘/⊘ par test et une barre M of N. En cas d'échec,
-    le caller imprime le résumé/traceback contenu dans ``result.output``.
-    """
-    from rich.progress import (
-        BarColumn,
-        MofNCompleteColumn,
-        Progress,
-        SpinnerColumn,
-        TextColumn,
-        TimeElapsedColumn,
-    )
-
-    state: dict[str, Any] = {"done": 0, "task_id": None}
-
-    # Mode machine : la barre et les verdicts partent sur stdout et
-    # rendraient le document JSON illisible. On lance les tests sans rien
-    # afficher — mesuré, sans cela la sortie commence par « ℹ Validation… ».
-    if quiet:
-        return check_lab(lab, target=target)
-
-    with Progress(
-        SpinnerColumn(),
-        TextColumn("[bold]{task.description}"),
-        BarColumn(bar_width=30),
-        MofNCompleteColumn(),
-        TimeElapsedColumn(),
-        transient=False,
-        console=console,
-    ) as progress:
-        task = progress.add_task("Collecte des tests…", total=None)
-        state["task_id"] = task
-
-        def on_event(event: dict[str, Any]) -> None:
-            etype = event.get("type")
-            if etype == "collected":
-                total = event.get("total", 0) or None
-                progress.update(
-                    task,
-                    description=_("progress_tests_running", lab_id=lab.id),
-                    total=total,
-                )
-            elif etype == "verdict":
-                nodeid = event.get("nodeid", "?")
-                verdict = event.get("verdict", "?")
-                # Raccourcit le nodeid pour l'affichage : ne garde que test_xxx::test_yyy
-                short = nodeid.rsplit("/", 1)[-1]
-                state["done"] += 1
-                progress.update(task, completed=state["done"])
-                if verdict == "PASSED":
-                    progress.console.print(f"  [green]✔[/green] {short}")
-                elif verdict in ("FAILED", "ERROR"):
-                    progress.console.print(f"  [red]✘ {short}  [dim]({verdict})[/dim][/red]")
-                elif verdict == "SKIPPED":
-                    progress.console.print(f"  [dim]⊘ {short}  (skipped)[/dim]")
-                else:
-                    progress.console.print(f"  [yellow]·[/yellow] {short}  [dim]({verdict})[/dim]")
-            # Les autres lignes (log/header/traceback) sont gardées dans
-            # result.output et imprimées seulement si le check échoue.
-
-        result = check_lab(lab, target=target, on_event=on_event)
-        progress.update(task, description=_("progress_tests_done", lab_id=lab.id))
-
-    return result
-
-
-def _run_check(
-    root: Path, lab: LabDefinition, target: str | None = None, *, quiet: bool = False,
-) -> tuple[CheckResult, int, int]:
-    """Lance les tests, enregistre le résultat, retourne (result, score, max_score).
-
-    ``target`` (option ``--target``) l'emporte sur la target active de la
-    session ; à défaut, la target ``default`` du lab s'applique.
-    """
-    # Même logique que pour --target : on refuse de NOTER ce qui n'a pas pu
-    # tourner. Un lab VM sans infrastructure n'est pas un échec de l'apprenant
-    # et ne doit pas lui coûter un 0/100 dans son historique. pytest tourne en
-    # sous-processus, donc l'erreur du conftest ne remonterait pas jusqu'ici :
-    # il faut vérifier AVANT.
-    if lab.runtime.type.value in ("vm", "kvm", "incus"):
-        from .discovery.repo import read_repo_metadata
-        from .infra.inventory import build_inventory, read_terraform_outputs
-
-        repo_meta = read_repo_metadata(root)
-        if repo_meta is not None:
-            try:
-                build_inventory(
-                    repo_meta,
-                    terraform_outputs=read_terraform_outputs(repo_meta),
-                )
-            except InfraNotProvisioned:
-                error(_("infra_not_provisioned"))
-                raise typer.Exit(2) from None
-            except ProviderUnresolved as exc:
-                # Un dépôt qui déclare plusieurs providers sans qu'aucun ne
-                # soit actif : lire les outputs Terraform est impossible, mais
-                # ce n'est pas une faute de l'apprenant. Sans ce garde-fou, la
-                # traceback remontait telle quelle depuis inventory.py.
-                if not exc.candidates:
-                    error(_("provider_none_declared"))
-                else:
-                    error(_("provider_required",
-                            candidates=", ".join(exc.candidates),
-                            first=exc.candidates[0]))
-                raise typer.Exit(2) from None
-
-    # Un --target explicite et inconnu est une ERREUR : on sort avant de
-    # lancer quoi que ce soit, sinon une faute de frappe enregistrerait un
-    # 0/100 dans l'historique de l'apprenant.
-    if target is not None and lab.runtime.target(target) is None:
-        declared = ", ".join(t.name for t in lab.runtime.targets) or "—"
-        error(_("unknown_target", target=target, declared=declared))
-        raise typer.Exit(1)
-
-    # À défaut, la target de session. Elle vaut pour TOUS les labs du dépôt :
-    # si celui-ci ne la déclare pas (lab shell, lab mono-target), on l'ignore
-    # simplement — ce n'est pas une erreur de l'apprenant.
-    if target is None:
-        session_target = read_context(root).active_target
-        if session_target and lab.runtime.target(session_target) is not None:
-            target = session_target
-
-    # Le verrou couvre les services ET pytest. Les tests pilotent la machine
-    # du lab (ou ses conteneurs) : deux validations concurrentes se marchent
-    # dessus, et la seconde note un état que la première est en train de
-    # changer. Il est pris ICI, pas au début : tout ce qui précède ne fait que
-    # lire, et refuser une faute de frappe sur `--target` pour cause de verrou
-    # serait absurde.
-    with _verrou(root, "check"):
-        return _valider(root, lab, target, quiet=quiet)
-
-
-def _valider(
-    root: Path, lab: LabDefinition, target: str | None, *, quiet: bool,
-) -> tuple[CheckResult, int, int]:
-    """Joue les tests et enregistre la note. Appelé sous verrou."""
-    # Les services conteneurisés (émulateur cloud, base…) doivent être debout
-    # avant que pytest ne s'exécute : les tests pilotent l'API qu'ils exposent.
-    _ensure_services(lab, root)
-
-    if not quiet:
-        info(_("validating", lab_id=lab.id))
-    result = _run_check_with_progress(lab, target, quiet=quiet)
-    if not result.ok and not quiet:
-        # En cas d'échec, dump l'output brut (tracebacks, summary pytest)
-        # pour que l'apprenant voie les erreurs détaillées.
-        #
-        # « and not quiet » : sans lui, la sortie pytest précédait le document
-        # JSON sur stdout dès qu'un test échouait, et le flux n'était plus
-        # analysable. Le cas le plus fréquent en usage réel, et le plus facile
-        # à manquer : un lab qui passe n'emprunte jamais cette branche.
-        # L'appelant en mode machine retrouve ce texte dans check.output.
-        console.print(result.output)
-
-    evaluation = evaluate_lab(root, lab, result)
-    if quiet:
-        # Mode machine : le tableau Rich et le message de confirmation
-        # pollueraient le document JSON. Le résultat est tout de même
-        # enregistré, comme dans le mode normal.
-        return result, evaluation.score, evaluation.max_score
-    print_check_result(
-        lab.id,
-        result.passed,
-        result.total,
-        evaluation.max_score,
-        evaluation.score,
-        evaluation.hints_used,
-        evaluation.hints_cost,
-    )
-    info(_("check_result_saved", score=evaluation.score, max_score=evaluation.max_score))
-    return result, evaluation.score, evaluation.max_score
-
-
-# ── check ─────────────────────────────────────────────────────────────────────
-
-@app.command("check", help=_("cmd_check_help"))
-def check(
-    lab_id: Annotated[str | None, typer.Argument(help=_("cmd_check_arg"), autocompletion=_complete_lab_id)] = None,
-    target: Annotated[str | None, typer.Option("--target", "-t",
-        help=_("opt_check_target"))] = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lab = _resolve_lab(root, lab_id, _lang(root))
-    try:
-        result, score, max_score = _run_check(root, lab, target, quiet=as_json)
-    except Interrupted as exc:
-        _interrompu(exc, f"dsoxlab check {lab.id}")
-    if as_json:
-        # La sortie brute de pytest est conservée : c'est là que l'appelant
-        # trouve le détail des échecs, qu'aucun compteur ne résume.
-        machine.emit({
-            "lab": machine.lab_dict(lab),
-            "check": {
-                "ok": result.ok,
-                "passed": result.passed,
-                "total": result.total,
-                "score": score,
-                "max_score": max_score,
-                "output": result.output,
-            },
-        })
-        if not result.ok:
-            raise typer.Exit(1)
-        return
-    if result.ok:
-        success(_("all_tests_passed"))
-        info(_("check_tip_submit"))
-    else:
-        error(_("tests_failed"))
-        raise typer.Exit(1)
-
-
-# ── submit ────────────────────────────────────────────────────────────────────
-
-@app.command("submit", help=_("cmd_submit_help"))
-def submit(
-    lab_id: Annotated[str | None, typer.Argument(help=_("cmd_submit_arg"), autocompletion=_complete_lab_id)] = None,
-    target: Annotated[str | None, typer.Option("--target", "-t",
-        help=_("opt_check_target"))] = None,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lab = _resolve_lab(root, lab_id, _lang(root))
-    try:
-        result, score, max_score = _run_check(root, lab, target)
-    except Interrupted as exc:
-        _interrompu(exc, f"dsoxlab submit {lab.id}")
-
-    if result.ok:
-        success(_("submit_success", score=score, max_score=max_score))
-    else:
-        info(_("submit_partial", passed=result.passed, total=result.total, score=score, max_score=max_score))
-
-    # Un lab qui déclare un seuil est un examen blanc, et un examen rend un
-    # verdict. Sans lui, un apprenant qui rendait 40/100 sur un mock RHCSA ne
-    # lisait nulle part qu'il avait échoué : la note s'affichait, jamais son
-    # sens. Un lab ordinaire n'en déclare pas et n'affiche donc rien.
-    from .services.progress_service import exam_percentage, exam_verdict
-
-    verdict = exam_verdict(score, max_score, lab.exam_passing_score)
-    if verdict is not None:
-        cle = "exam_passed" if verdict else "exam_failed"
-        rendu = success if verdict else error
-        rendu(_(
-            cle,
-            pct=exam_percentage(score, max_score),
-            threshold=lab.exam_passing_score,
-        ))
-
-    set_active_lab(root, None)
-    console.print()
-    # CTA "tape exit" uniquement si on est dans le sous-shell ouvert
-    # par ``dsoxlab run`` (cas runtime shell). Sur runtime vm,
-    # l'apprenant est revenu sur son poste local — pas de sous-shell
-    # à fermer, donc le message serait trompeur.
-    if os.environ.get("DSOXLAB_LAB_SESSION"):
-        console.print(_("submit_exit_cta"))
-    else:
-        console.print(_("submit_done"))
-
-
-# ── scores ────────────────────────────────────────────────────────────────────
-
-@app.command("scores", help=_("cmd_scores_help"))
-def scores(
-    lab_home: LabHomeOption = None,
-    section: Annotated[str | None, typer.Option("--section", "-s", help=_("opt_section"))] = None,
-    lab_id: Annotated[str | None, typer.Option("--lab", "-l", help=_("opt_filter_lab"))] = None,
-    top: Annotated[int, typer.Option("--top", help=_("opt_top"))] = 20,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    root = _root(lab_home)
-    ctx = read_context(root)
-    effective_section = section or ctx.section
-    results = get_results(root, lab_id=lab_id, section=effective_section, limit=top)
-    # Les seuils vivent dans le catalogue, les notes dans la base : le verdict
-    # d'un examen demande les deux. On ne balaie que pour les labs affichés.
-    affiches = {r["lab_id"] for r in results}
-    seuils = {
-        lab.id: lab.exam_passing_score
-        for lab in _catalogue(root, _lang(root), quiet=as_json)
-        if lab.exam_passing_score and lab.id in affiches
-    }
-    if as_json:
-        machine.emit({
-            "results": [machine.score_dict(r, seuils.get(r["lab_id"])) for r in results],
-            "count": len(results),
-        })
-        return
-    print_scores_table(results, seuils)
-
-
-# ── progress ──────────────────────────────────────────────────────────────────
-
-@app.command("progress", help=_("cmd_progress_help"))
-def progress(
-    lab_home: LabHomeOption = None,
-    section: Annotated[str | None, typer.Option("--section", "-s", help=_("opt_section"))] = None,
-    level: Annotated[str | None, typer.Option("--level", "-l", help=_("opt_level"))] = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    root = _root(lab_home)
-    ctx = read_context(root)
-    lang = _lang(root)
-
-    effective_section = section or ctx.section
-    effective_level = level or ctx.level
-
-    labs = _catalogue(root, lang, quiet=as_json)
-    if effective_section:
-        labs = [lab for lab in labs if lab.section == effective_section]
-    if effective_level:
-        labs = [lab for lab in labs if lab.level == effective_level]
-
-    # Sort by bloc then bloc_order for a coherent display
-    labs = sorted(labs, key=lambda lab: (lab.bloc, lab.bloc_order, lab.id))
-
-    lab_ids = [lab.id for lab in labs]
-    scores_data = get_best_scores(root, lab_ids)
-    if as_json:
-        faits = [i for i in lab_ids if i in scores_data]
-        machine.emit({
-            "labs": [machine.lab_dict(lab, scores_data.get(lab.id)) for lab in labs],
-            "summary": {
-                "total": len(labs),
-                "attempted": len(faits),
-                "points": sum(scores_data[i][0] for i in faits),
-                "max_points": sum(scores_data[i][1] for i in faits),
-            },
-        })
-        return
-    print_progress_table(labs, scores_data)
-
-
-# ── next ──────────────────────────────────────────────────────────────────────
-
-@app.command("next", help=_("cmd_next_help"))
-def next_lab(
-    lab_home: LabHomeOption = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    root = _root(lab_home)
-    ctx = read_context(root)
-    lang = _lang(root)
-
-    # Erreur dure, en `--json` comme ailleurs : sans contexte il n'y a rien à
-    # décrire. La cause part sur stderr, la sortie standard reste vide, et le
-    # code de retour ne change pas.
-    if not ctx.section:
-        error(_("next_no_context"))
-        raise typer.Exit(1)
-
-    labs = _catalogue(root, lang, quiet=as_json)
-    if ctx.section:
-        labs = [lab for lab in labs if lab.section == ctx.section]
-    if ctx.level:
-        labs = [lab for lab in labs if lab.level == ctx.level]
-
-    scores_data = get_best_scores(root, [lab.id for lab in labs])
-
-    upcoming = next_pending_lab(labs, scores_data)
-    if as_json:
-        # `next` et `all_done` disent deux choses différentes : un contexte
-        # sans lab du tout rendrait aussi `next: null`, et l'appelant fêterait
-        # une section terminée qui est en fait vide.
-        machine.emit({
-            "context": {"section": ctx.section, "level": ctx.level or None},
-            "next": None if upcoming is None
-            else machine.lab_dict(upcoming, scores_data.get(upcoming.id)),
-            "all_done": upcoming is None and bool(labs),
-            "remaining": sum(1 for lab in labs if lab.id not in scores_data),
-        })
-        return
-    if upcoming is None:
-        success(_("next_all_done"))
-        return
-    success(_("next_suggestion", lab_id=upcoming.id, title=upcoming.title))
-
-
-# ── reset ─────────────────────────────────────────────────────────────────────
-
-@app.command("reset", help=_("cmd_reset_help"))
-def reset(
-    ctx: typer.Context,
-    lab_id: Annotated[str, typer.Argument(help=_("cmd_reset_arg"), autocompletion=_complete_lab_id)],
-    target: Annotated[str | None, typer.Option("--target", "-t",
-        help=_("opt_run_target"))] = None,
-    lab_home: LabHomeOption = None,
-) -> None:
-    root = _root(lab_home)
-    lang = _lang(root)
-    try:
-        lab = _lab(root, lab_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    ctx.call_on_close(_verrou(root, "reset").release)
-    info(_("resetting", lab_id=lab.id))
-    try:
-        if lab.runtime.type.value in ("vm", "kvm", "incus"):
-            _run_ansible_with_progress(
-                lab.path / "cleanup.yaml",
-                lambda cb: reset_lab(lab, target_name=target, on_event=cb),
-            )
-        else:
-            reset_lab(lab, target_name=target)
-        reset_hints(root, lab.id)
-        success(_("lab_reset"))
-    except Interrupted as exc:
-        _interrompu(exc, f"dsoxlab reset {lab.id}")
-    except RuntimeError as exc:
-        error(str(exc))
-        raise typer.Exit(2) from None
-
-
-# ── clean ─────────────────────────────────────────────────────────────────────
-
-@app.command("clean", help=_("cmd_clean_help"))
-def clean(
-    ctx: typer.Context,
-    lab_id: Annotated[str, typer.Argument(help=_("cmd_clean_arg"), autocompletion=_complete_lab_id)],
-    target: Annotated[str | None, typer.Option("--target", "-t",
-        help=_("opt_run_target"))] = None,
-    lab_home: LabHomeOption = None,
-    yes: Annotated[bool, typer.Option("--yes", "-y", help=_("opt_yes"))] = False,
-) -> None:
-    root = _root(lab_home)
-    lang = _lang(root)
-    try:
-        lab = _lab(root, lab_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    if not yes:
-        typer.confirm(_("confirm_clean", lab_id=lab.id), abort=True)
-
-    # Après la confirmation, jamais avant : tenir le verrou pendant qu'on
-    # attend une réponse au clavier bloquerait l'autre terminal sur une
-    # question que personne ne voit.
-    ctx.call_on_close(_verrou(root, "clean").release)
-    info(_("cleaning", lab_id=lab.id))
-    try:
-        if lab.runtime.type.value in ("vm", "kvm", "incus"):
-            _run_ansible_with_progress(
-                lab.path / "cleanup.yaml",
-                lambda cb: clean_lab(lab, target_name=target, on_event=cb),
-            )
-        else:
-            clean_lab(lab, target_name=target)
-        _stop_services(lab, root)
-        success(_("clean_done"))
-    except Interrupted as exc:
-        _interrompu(exc, f"dsoxlab clean {lab.id}")
-    except RuntimeError as exc:
-        error(str(exc))
-        raise typer.Exit(2) from None
-
-
-# ── validate-structure ────────────────────────────────────────────────────────
-
-#: Les familles de contrôle de ``validate-structure``, dans l'ordre où elles
-#: sont jouées. Elles figurent **toutes** dans ``counts``, à zéro s'il le faut :
-#: un tableau de bord qui n'aurait pas la clé ne saurait pas si la famille est
-#: saine ou si cette version de l'outil ne la connaît pas.
-_FAMILLES_ANOMALIES = (
-    "contract", "unknown_key", "structure", "content", "doc_url", "metadata",
-)
-
-
-def _compter(anomalies: list[dict[str, Any]]) -> dict[str, int]:
-    """Le nombre d'anomalies par famille, toutes familles présentes."""
-    compte = dict.fromkeys(_FAMILLES_ANOMALIES, 0)
-    for anomalie in anomalies:
-        compte[str(anomalie["kind"])] += 1
-    return compte
-
-
-@app.command("validate-structure", help=_("cmd_validate_help"))
-def validate_structure_cmd(
-    lab_home: LabHomeOption = None,
-    check_urls: Annotated[bool, typer.Option(
-        "--check-urls", help=_("opt_check_urls"),
-    )] = False,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    from .discovery.repo import read_repo_metadata
-    from .discovery.scanner import discover_labs
-    from .validators.content import (
-        ContentIssue,
-        check_doc_url,
-        validate_internal_links,
-        validate_language_parity,
-        validate_scoring,
-        validate_solutions_encrypted,
-        validate_targets,
-    )
-    from .validators.contract import validate_schema_versions, validate_unknown_keys
-
-    root = _root(lab_home)
-
-    def _rendu(chemin: Path) -> Path:
-        try:
-            return chemin.relative_to(root)
-        except ValueError:
-            return chemin
-
-    # Le document machine se construit au fil des contrôles, dans le même ordre
-    # que l'affichage : une seule passe, deux rendus, et donc aucun risque que
-    # l'un rapporte une anomalie que l'autre tait.
-    documents: list[dict[str, Any]] = []
-
-    def _rendre(entete: str, lignes: list[str]) -> None:
-        """Affiche un groupe d'anomalies, sauf en mode machine."""
-        if as_json or not lignes:
-            return
-        console.print(_(entete))
-        for ligne in lignes:
-            console.print(ligne)
-
-    # D'ABORD, et à la source : un `schema_version` illisible ou trop récent
-    # empêche le fichier d'être découvert, donc TOUS les contrôles suivants
-    # l'ignoreraient — c'est le trou connu du validator, qui n'itère que sur ce
-    # qui a déjà été chargé. Ce contrôle-ci relit les fichiers du disque.
-    contract = validate_schema_versions(root)
-    documents += [
-        machine.issue_dict("contract", a.key, a.params, path=a.path)
-        for a in contract.issues
-    ]
-    _rendre("contract_issues_header", [
-        f"  [red]✘[/red] {_rendu(a.path)}: {_(a.key, **a.params)}"
-        for a in contract.issues
-    ])
-    # Le meta.yml décrit tout le catalogue : illisible, il rend chaque
-    # contrôle suivant douteux, et la découverte lèverait de toute façon.
-    # Un lab isolé, lui, n'empêche pas de valider les 283 autres.
-    if not contract.ok and contract.meta_is_unreadable:
-        if as_json:
-            # Même forme de document que la sortie complète : un appelant ne
-            # doit pas avoir deux structures à gérer selon l'endroit où la
-            # validation s'est arrêtée.
-            machine.emit({
-                "ok": False,
-                "labs_checked": 0,
-                "doc_urls_checked": False,
-                "issues": documents,
-                "counts": _compter(documents),
-            })
-            raise typer.Exit(1)
-        error(_("labs_have_issues"))
-        raise typer.Exit(1)
-
-    # Ensuite, toujours à la source : les clés que le moteur n'ira jamais lire.
-    # Le parseur les ignore et continuera de le faire — c'est une garantie de
-    # la v1 — mais « toléré » n'est pas « voulu » : onze labs d'examen ont posé
-    # un seuil de réussite que personne ne lisait, sans que rien ne le dise.
-    unknown = validate_unknown_keys(root)
-    documents += [
-        machine.issue_dict("unknown_key", a.key, a.params, path=a.path)
-        for a in unknown.issues
-    ]
-    _rendre("unknown_keys_header", [
-        f"  [red]✘[/red] {_rendu(a.path)}: {_(a.key, **a.params)}"
-        for a in unknown.issues
-    ])
-
-    structure_reports = validate_all_structure(root)
-    metadata_reports = validate_all_metadata(root)
-
-    documents += [
-        machine.issue_dict("structure", i.key, i.params, lab=r.lab_id, path=i.path)
-        for r in structure_reports for i in r.issues
-    ]
-    if not as_json:
-        print_structure_reports(structure_reports)
-
-    # Contrôles de contenu : locaux, donc jouables hors ligne et par défaut.
-    # Un lien mort ou une solution en clair ne casse aucun test fonctionnel,
-    # c'est bien pourquoi rien ne les attrapait.
-    labs = discover_labs(root)
-    # Les cibles vm se valident contre infra.hosts : un FQDN inconnu ne se
-    # voyait qu'au run, sur la machine de l'apprenant, après provisionnement.
-    try:
-        repo_meta = read_repo_metadata(root)
-        host_names = {h.name for h in repo_meta.infra.hosts} if repo_meta else set()
-    except Exception:  # noqa: BLE001 - meta.yml illisible : les autres contrôles restent utiles
-        host_names = set()
-
-    content_issues: list[tuple[str, Path, ContentIssue]] = []
-    for lab in labs:
-        rapports = [
-            validate_internal_links(lab),
-            validate_scoring(lab),
-            validate_language_parity(lab),
-            validate_targets(lab, host_names),
-        ]
-        # « solution/<chemin du lab depuis labs/> » : convention respectée par
-        # les dépôts qui tiennent leurs corrigés hors des labs. Absent = pas
-        # de contrôle, ce n'est pas une faute.
-        try:
-            relatif = lab.path.relative_to(root / "labs")
-        except ValueError:
-            relatif = None
-        if relatif is not None:
-            rapports.append(
-                validate_solutions_encrypted(lab, root / "solution" / relatif)
-            )
-        for rapport in rapports:
-            for souci in rapport.issues:
-                content_issues.append((lab.id, souci.path, souci))
-
-    documents += [
-        machine.issue_dict("content", s.key, s.params, lab=lab_id, path=chemin)
-        for lab_id, chemin, s in content_issues
-    ]
-    _rendre("content_issues_header", [
-        f"  [red]✘[/red] {lab_id} — {_rendu(chemin)}: {_(s.key, **s.params)}"
-        for lab_id, chemin, s in content_issues
-    ])
-
-    url_issues: list[tuple[str, str, ContentIssue]] = []
-    if check_urls:
-        # Sur stdout : tue en mode machine, où un « ℹ » suffit à rendre le
-        # document illisible.
-        if not as_json:
-            info(_("checking_doc_urls", count=len(labs)))
-        for lab in labs:
-            injoignable = check_doc_url(lab)
-            if injoignable is not None:
-                url_issues.append((lab.id, lab.doc_url, injoignable))
-        documents += [
-            # L'URL entre dans les paramètres : c'est le fait que l'appelant
-            # veut, et il n'a pas à relire le catalogue pour l'obtenir.
-            machine.issue_dict(
-                "doc_url", r.key, {**r.params, "url": url}, lab=lab_id, path=r.path,
-            )
-            for lab_id, url, r in url_issues
-        ]
-        _rendre("doc_url_issues_header", [
-            f"  [red]✘[/red] {lab_id} — {url} — {_(r.key, **r.params)}"
-            for lab_id, url, r in url_issues
-        ])
-
-    issues = [r for r in metadata_reports if not r.ok]
-    documents += [
-        machine.issue_dict("metadata", i.key, i.params, lab=r.lab_id, field=i.field)
-        for r in issues for i in r.issues
-    ]
-    _rendre("metadata_issues_header", [
-        f"  [red]✘[/red] {r.lab_id} — {i.field}: {_(i.key, **i.params)}"
-        for r in issues for i in r.issues
-    ])
-
-    all_ok = (
-        contract.ok
-        and unknown.ok
-        and all(r.ok for r in structure_reports)
-        and not issues
-        and not content_issues
-        and not url_issues
-    )
-    if as_json:
-        machine.emit({
-            "ok": all_ok,
-            "labs_checked": len(labs),
-            # `--check-urls` est le seul contrôle réseau, et le seul qui ne
-            # tourne pas par défaut : sans ce champ, un appelant ne peut pas
-            # distinguer « aucune URL morte » de « les URL n'ont pas été
-            # regardées ».
-            "doc_urls_checked": check_urls,
-            "issues": documents,
-            "counts": _compter(documents),
-        })
-        # Le verdict et le code de retour sont ceux du mode terminal : seule la
-        # forme de la sortie change.
-        if not all_ok:
-            raise typer.Exit(1)
-        return
-    if all_ok:
-        success(_("all_labs_valid"))
-    else:
-        error(_("labs_have_issues"))
-        raise typer.Exit(1)
-
-
-# ── doctor ────────────────────────────────────────────────────────────────────
-
-@app.command("doctor", help=_("cmd_doctor_help"))
-def doctor(
-    lab_home: LabHomeOption = None,
-    fix: Annotated[bool, typer.Option("--fix", help=_("opt_fix"))] = False,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    root = _root(lab_home)
-
-    # `--fix` joue des commandes système (`apt`, `systemctl`, `usermod`) dont la
-    # sortie va sur la sortie standard, qu'aucune option ne détourne. Les deux
-    # options ensemble produiraient donc un document précédé de la sortie d'apt,
-    # c'est-à-dire un flux qu'aucun appelant ne peut lire. On le dit plutôt que
-    # de rendre du JSON cassé.
-    if as_json and fix:
-        error(_("doctor_json_sans_fix"))
-        raise typer.Exit(1)
-
-    # Le meta.yml décide de ce qui est bloquant ici : un dépôt sans lab `vm`
-    # n'a besoin d'aucun hyperviseur, et un dépôt qui a choisi son provider
-    # n'a pas besoin des autres. Son absence n'est pas une erreur : on
-    # diagnostique alors le strict socle commun.
-    try:
-        repo_meta = _read_repo(root)
-    except typer.Exit:
-        # Un meta.yml illisible est justement ce qu'un diagnostic doit
-        # pouvoir rapporter : l'erreur est déjà affichée, on poursuit sur
-        # le socle commun plutôt que de sortir sans rien dire d'autre.
-        repo_meta = None
-    report = collect_checks(root, repo_meta)
-
-    if as_json:
-        # Le code de retour ne change pas : `doctor` sort en 0 même quand un
-        # contrôle échoue, et le verdict se lit dans `ok`. Un `--json` qui
-        # inventerait un code non nul ferait diverger les deux modes.
-        machine.emit(machine.doctor_dict(report))
-        return
-
-    print_doctor(report)
-
-    if fix:
-        failing = [(c.label, c.fix) for c in report.fixable() if c.fix]
-        if not failing:
-            info(_("fix_nothing"))
-            return
-
-        # Pr\u00e9-conditions sudo : si au moins un fix d\u00e9marre par "sudo", on
-        # v\u00e9rifie que l'env est compatible avant d'attaquer (TTY pour
-        # taper le password, sudo dispo, et id\u00e9alement on cache le
-        # password une seule fois pour toute la cascade).
-        sudo_fixes = [c for _, c in failing if c.strip().startswith("sudo ")]
-        if sudo_fixes:
-            if not sys.stdin.isatty():
-                error(_("fix_needs_tty"))
-                raise typer.Exit(1)
-            if shutil.which("sudo") is None:
-                error(_("fix_no_sudo"))
-                raise typer.Exit(1)
-
-            # Pr\u00e9-cache les credentials sudo : un seul prompt password
-            # pour toute la cascade. Sans ce sudo -v, l'apprenant
-            # pourrait avoir \u00e0 retaper son password \u00e0 chaque commande
-            # (si sudo timestamp_timeout=0 ou si la cascade d\u00e9passe 5min).
-            info(_("fix_sudo_preauth", count=len(sudo_fixes)))
-            # check=False : un mot de passe refusé ou un Ctrl-C sur le prompt
-            # sudo est une réponse de l'utilisateur, pas une panne. On la
-            # traduit en message et en code de sortie, pas en trace Python.
-            preauth = subprocess.run(["sudo", "-v"], check=False)
-            if preauth.returncode != 0:
-                error(_("fix_sudo_failed"))
-                raise typer.Exit(1)
-
-        info(_("fix_count", count=len(failing)))
-        for label, fix_cmd in failing:
-            info(f"[bold]{label}[/bold] \u2192 {fix_cmd}")
-            # check=False : `--fix` joue une CASCADE. Un correctif en \u00e9chec
-            # doit \u00eatre signal\u00e9 puis laisser tourner les suivants ; lever ici
-            # abandonnerait les r\u00e9parations restantes sur la premi\u00e8re qui rate.
-            result = subprocess.run(fix_cmd, shell=True, text=True, check=False)  # noqa: S602
-            if result.returncode == 0:
-                success(_("fix_success", label=label))
-            else:
-                error(_("fix_failure", label=label, code=result.returncode))
-        info(_("fix_rerun"))
-
-
-# ── provision / destroy / ssh / status ────────────────────────────────────────
-
-
-def _undefine_command(orphans: dict[str, str]) -> str:
-    """La commande exacte qui retire des machines restées sur l'hyperviseur.
-
-    Rendue copiable telle quelle : dire « supprime-les » sans donner le geste
-    laisse l'apprenant chercher ``virsh undefine``, qu'aucune page du parcours
-    ne lui a montré. ``sudo`` sans ``-n``, ici : c'est un humain qui la tape,
-    un mot de passe demandé au terminal ne pose aucun problème.
-    """
-    return "; ".join(
-        f"sudo virsh undefine --nvram {domain}" for domain in sorted(orphans.values())
-    )
-
-
-def _diagnostic_message(hote: dict[str, Any]) -> str:
-    """La phrase qui nomme la cause d'un hôte muet, et le geste qui la corrige.
-
-    Une cause, un message, une action. L'ancien texte en proposait deux à la
-    fois — « cloud-init tourne peut-être encore, ou alors reprovisionne » — pour
-    tous les hôtes et toutes les pannes : l'apprenant devait trancher lui-même
-    entre deux conseils dont l'un coûtait une infrastructure entière.
-    """
-    cause = hote["cause"]
-    domaine = hote.get("domain") or hote["fqdn"]
-    if cause == host_diagnosis.CAUSE_DOMAIN_ABSENT:
-        return _("status_cause_domain_absent", host=hote["fqdn"])
-    if cause == host_diagnosis.CAUSE_DOMAIN_NOT_RUNNING:
-        return _("status_cause_domain_not_running",
-                 domain=domaine, state=hote.get("domain_state") or "?")
-    if cause == host_diagnosis.CAUSE_DOMAIN_NO_LEASE:
-        return _("status_cause_domain_no_lease", domain=domaine)
-    if cause == host_diagnosis.CAUSE_BOOTING:
-        return _("status_cause_booting", domain=domaine)
-    if cause == host_diagnosis.CAUSE_SSH_REFUSED:
-        return _("status_cause_ssh_refused", ip=hote["ip"])
-    if cause == host_diagnosis.CAUSE_UNREACHABLE:
-        return _("status_cause_unreachable", ip=hote["ip"])
-    if cause == host_diagnosis.CAUSE_SSH_TIMEOUT:
-        return _("status_cause_ssh_timeout", ip=hote["ip"])
-    if cause == host_diagnosis.CAUSE_SSH_DENIED:
-        return _("status_cause_ssh_denied", ip=hote["ip"])
-    return _("status_cause_unknown", reason=hote.get("reason") or "?")
-
-
-@app.command("provision", help=_("cmd_provision_help"))
-def provision(
-    ctx: typer.Context,
-    host: Annotated[list[str] | None, typer.Option(
-        "--host",
-        help=_("opt_provision_host"),
-    )] = None,
-    lab_home: LabHomeOption = None,
-) -> None:
-    """Lance terraform apply sur le provider courant avec progress bar."""
-    from .infra import terraform as tf
-    from .infra.terraform import ProviderNotImplemented, TerraformNotInstalled, host_targets
-
-    root = _root(lab_home)
-    # Le verrou est pris avant même de lire le contrat : ce qui suit interroge
-    # l'hyperviseur, puis écrit le state Terraform. Un `destroy` lancé dans un
-    # autre terminal pendant le scan des machines orphelines rendrait ce scan
-    # faux au moment où on s'en sert.
-    ctx.call_on_close(_verrou(root, "provision").release)
-    repo_meta = _read_repo(root)
-    if repo_meta is None:
-        error(_("provision_no_meta", root=root))
-        raise typer.Exit(1)
-
-    provider = _require_provider(repo_meta)
-
-    # Garde-fou cohabitation : incus et KVM partagent le nom de réseau et le
-    # subnet du lab → ils ne peuvent pas coexister. Si un AUTRE provider a
-    # encore de l'infra debout, on guide l'apprenant (destroy) plutôt que de
-    # le laisser buter sur « Network is already in use ».
-    conflicts = tf.other_active_providers(repo_meta)
-    if conflicts:
-        error(_(
-            "provision_provider_conflict",
-            current=provider,
-            others=", ".join(conflicts),
-            other=conflicts[0],
-        ))
-        raise typer.Exit(5)
-
-    # Garde-fou « machines fantômes » : un provisionnement interrompu après la
-    # définition d'un domaine le laisse sur l'hyperviseur sans jamais l'inscrire
-    # au state. Terraform ne le voit donc pas, et tout apply suivant meurt sur
-    # « domain already exists », une phrase qui ne dit pas quoi faire. On nomme
-    # les machines et la commande qui les retire, avant de perdre une minute
-    # d'init et d'apply.
-    scan = tf.find_orphan_domains(repo_meta)
-    if scan.reason:
-        warn(_("orphan_check_skipped", error=scan.reason))
-    if scan.orphans:
-        error(_("provision_orphan_domains", hosts=", ".join(sorted(scan.orphans))))
-        info(_("provision_orphan_fix", cmd=_undefine_command(scan.orphans)))
-        raise typer.Exit(5)
-
-    info(_("provision_starting", provider=provider))
-
-    # Garde-fou : la clé SSH du repo doit exister avant tout provision.
-    # Sans elle, le keypair cloud serait créé avec une clé publique
-    # vide → VMs inaccessibles. Cas vécu sur Outscale lors d'un test
-    # initial.
-    ssh_key = repo_meta.path / "ssh" / "id_ed25519"
-    if not ssh_key.is_file():
-        error(_("provision_no_ssh_key", path=ssh_key))
-        raise typer.Exit(1)
-
-    # Construit la liste des targets Terraform pour les --host demandés.
-    targets: list[str] = []
-    if host:
-        known = {h.name for h in repo_meta.infra.hosts}
-        for fqdn in host:
-            if fqdn not in known:
-                error(_("host_unknown", fqdn=fqdn, known=", ".join(sorted(known))))
-                raise typer.Exit(1)
-            try:
-                targets.extend(host_targets(provider, fqdn))
-            except NotImplementedError as exc:
-                error(str(exc))
-                raise typer.Exit(2) from None
-        info(_("terraform_target", hosts=", ".join(host), count=len(targets)))
-
-    try:
-        # Étape 1 : terraform init (peut télécharger ~50 MB de provider
-        # au premier run). Spinner pour ne pas laisser l'utilisateur
-        # croire que ça plante.
-        _run_terraform_init_with_spinner(
-            lambda cb: tf.init(repo_meta, on_event=cb)
-        )
-        # Étape 2 : terraform apply avec progress bar par ressource
-        result = _run_terraform_with_progress(
-            "provision",
-            lambda cb: tf.apply(
-                repo_meta, on_event=cb,
-                targets=targets or None, target_hosts=list(host) if host else None,
-            ),
-        )
-    except Interrupted as exc:
-        # AVANT le `except Exception` : une interruption n'est pas un échec de
-        # provisionnement, et sortir en 4 avec « provision failed » ferait
-        # chercher une panne là où l'apprenant a simplement appuyé sur Ctrl-C.
-        _interrompu(exc, "dsoxlab provision")
-    except ProviderNotImplemented as exc:
-        error(str(exc))
-        raise typer.Exit(2) from None
-    except TerraformNotInstalled as exc:
-        error(str(exc))
-        raise typer.Exit(3) from None
-    except Exception as exc:  # noqa: BLE001 — message utilisateur direct
-        error(_("provision_failed", error=str(exc)))
-        # Terraform est exact mais opaque pour qui découvre l'outil. Quand la
-        # cause est connue et le correctif tient en une ligne, on les donne
-        # plutôt que de laisser l'apprenant chercher : c'est là qu'il est
-        # bloqué, et c'est le seul moment où l'on peut l'affirmer sans risque
-        # de fausse alerte.
-        from .services.doctor import explique_echec_provision
-
-        connu = explique_echec_provision(str(exc))
-        if connu is not None:
-            explication, commande = connu
-            info(explication)
-            info(f"  {commande}")
-        raise typer.Exit(4) from None
-
-    # Étape 3 : attendre que les VMs soient réellement joignables (sshd +
-    # compte student + cloud-init terminé). Sans ça, le premier `dsoxlab run`
-    # échoue en « unreachable » car la VM boote encore.
-    from .infra.inventory import HostReadyTimeout, wait_for_hosts_ready
-
-    ready_hosts = sorted(result.hosts)
-    if ready_hosts:
-        from rich.progress import (
-            Progress,
-            SpinnerColumn,
-            TextColumn,
-            TimeElapsedColumn,
-        )
-
-        with Progress(
-            SpinnerColumn(),
-            TextColumn("[bold]{task.description}"),
-            TimeElapsedColumn(),
-            transient=True,
-            console=console,
-        ) as progress:
-            task = progress.add_task(_("provision_waiting_ssh"), total=None)
-
-            def _on_attempt(fqdn: str, attempt: int) -> None:
-                progress.update(
-                    task,
-                    description=_(
-                        "provision_waiting_ssh_host", host=fqdn, attempt=attempt
-                    ),
-                )
-
-            try:
-                with interruptible(Stage.HOSTS_WAIT):
-                    wait_for_hosts_ready(
-                        repo_meta, ready_hosts, on_attempt=_on_attempt
-                    )
-            except HostReadyTimeout as exc:
-                progress.stop()
-                warn(_("provision_ssh_timeout", error=str(exc)))
-            except Interrupted as exc:
-                # L'infrastructure existe déjà : c'est l'attente qui a été
-                # coupée, pas la création. Rejouer `provision` est idempotent
-                # et reprend exactement ici.
-                progress.stop()
-                _interrompu(exc, "dsoxlab provision")
-
-    # Le fragment SSH, écrit à CHAQUE provision et non seulement quand des
-    # machines viennent d'être créées : relancer un provision sur une infra
-    # déjà en place laissait sinon l'apprenant sans fragment, alors que c'est
-    # le moment où il en a besoin. Il doit refléter l'état courant, pas le
-    # delta du dernier terraform apply.
-    from .infra.inventory import (
-        build_inventory,
-        read_terraform_outputs,
-        ssh_config_include_present,
-        user_ssh_config_path,
-        write_ssh_config,
-        write_user_ssh_config,
-    )
-
-    fragment = user_ssh_config_path(repo_meta)
-    try:
-        inv = build_inventory(
-            repo_meta, terraform_outputs=read_terraform_outputs(repo_meta)
-        )
-        write_ssh_config(inv, repo_meta)
-        write_user_ssh_config(inv, repo_meta)
-    except (OSError, RuntimeError) as exc:
-        # Un fragment manquant ne doit pas faire échouer un provision réussi.
-        warn(_("ssh_fragment_failed", error=str(exc)))
-
-    if fragment.is_file():
-        if ssh_config_include_present():
-            info(_("ssh_fragment_written", path=fragment))
-        else:
-            warn(_("ssh_fragment_no_include", path=fragment))
-
-    success(_("provision_done", count=len(result.hosts)))
-    for fqdn, ip in sorted(result.hosts.items()):
-        info(f"  {fqdn} → {ip}")
-
-
-@app.command("destroy", help=_("cmd_destroy_help"))
-def destroy(
-    ctx: typer.Context,
-    host: Annotated[list[str] | None, typer.Option(
-        "--host",
-        help=_("opt_destroy_host"),
-    )] = None,
-    yes: Annotated[bool, typer.Option(
-        "--yes", "-y",
-        help=_("opt_yes"),
-    )] = False,
-    lab_home: LabHomeOption = None,
-) -> None:
-    """Lance terraform destroy sur le provider courant avec progress bar."""
-    from .infra import terraform as tf
-    from .infra.terraform import ProviderNotImplemented, TerraformNotInstalled, host_targets
-
-    root = _root(lab_home)
-    repo_meta = _read_repo(root)
-    if repo_meta is None:
-        error(_("provision_no_meta", root=root))
-        raise typer.Exit(1)
-
-    provider = _require_provider(repo_meta)
-
-    targets: list[str] = []
-    if host:
-        known = {h.name for h in repo_meta.infra.hosts}
-        for fqdn in host:
-            if fqdn not in known:
-                error(_("host_unknown", fqdn=fqdn, known=", ".join(sorted(known))))
-                raise typer.Exit(1)
-            try:
-                targets.extend(host_targets(provider, fqdn))
-            except NotImplementedError as exc:
-                error(str(exc))
-                raise typer.Exit(2) from None
-        info(_("terraform_target", hosts=", ".join(host), count=len(targets)))
-        # Mesuré le 2026-07-23 : terraform détruit la cible ET tout ce qui en
-        # dépend. Les volumes et disques cloud-init étant chaînés aux domaines,
-        # cibler un seul hôte emporte les autres (7 ressources détruites pour
-        # 1 demandée). On prévient plutôt que de laisser croire à un ciblage fin.
-        warn(_("destroy_host_not_isolated"))
-
-    # destroy est irréversible et ne prévenait pas : un « dsoxlab destroy »
-    # tapé dans le mauvais dépôt effaçait le parc sans un mot. --yes garde
-    # l'usage scripté (CI, procédure de récupération documentée).
-    if not yes:
-        typer.confirm(_("confirm_destroy", provider=provider), abort=True)
-
-    # Après la confirmation : tenir le verrou pendant l'attente d'une réponse
-    # au clavier bloquerait l'autre terminal sans rien protéger.
-    ctx.call_on_close(_verrou(root, "destroy").release)
-    info(_("destroy_starting", provider=provider))
-    _purge_snapshots_before_destroy(repo_meta)
-    try:
-        # init est rapide en destroy (provider déjà téléchargé) mais
-        # nécessaire si l'utilisateur a fait un upgrade dsoxlab entre temps
-        _run_terraform_init_with_spinner(
-            lambda cb: tf.init(repo_meta, on_event=cb)
-        )
-        _run_terraform_with_progress(
-            "destroy",
-            lambda cb: tf.destroy(
-                repo_meta, on_event=cb,
-                targets=targets or None, target_hosts=list(host) if host else None,
-            ),
-        )
-    except Interrupted as exc:
-        # Avant le `except Exception`, même raison qu'au provision : Terraform
-        # a écrit son state, ce qui reste debout est connu, et rejouer suffit.
-        _interrompu(exc, "dsoxlab destroy")
-    except ProviderNotImplemented as exc:
-        error(str(exc))
-        raise typer.Exit(2) from None
-    except TerraformNotInstalled as exc:
-        error(str(exc))
-        raise typer.Exit(3) from None
-    except Exception as exc:  # noqa: BLE001
-        error(_("destroy_failed", error=str(exc)))
-        raise typer.Exit(4) from None
-
-    # Terraform ne détruit que ce qu'il connaît. Un domaine défini puis jamais
-    # inscrit au state lui est invisible : il annonçait donc « infrastructure
-    # détruite » et sortait en 0 en laissant les machines debout, ce qui est le
-    # contraire de ce que la commande promet. On regarde l'hyperviseur.
-    if not _handle_orphans_after_destroy(repo_meta, assume_yes=yes):
-        raise typer.Exit(6)
-
-    # Le fragment SSH pointe désormais des machines mortes : le laisser
-    # enverrait l'apprenant vers des adresses recyclées, ce qui est pire que
-    # pas de configuration du tout.
-    from .infra.inventory import remove_user_ssh_config
-
-    if remove_user_ssh_config(repo_meta):
-        info(_("ssh_fragment_removed", repo=repo_meta.id))
-
-    success(_("destroy_done"))
-
-
-def _purge_snapshots_before_destroy(repo_meta: RepoMetadata) -> None:
-    """Retire les points de reprise **avant** que Terraform ne passe.
-
-    Un snapshot externe laisse un fichier de recouvrement que Terraform ne
-    connaît pas : il n'est dans aucun state, et le volume qu'il recouvre est
-    supprimé sous lui. Il faut donc le retirer tant que l'hyperviseur sait
-    encore qu'il existe — après l'``undefine``, la métadonnée du snapshot a
-    disparu avec le domaine, et le fichier devient introuvable autrement qu'à
-    la main.
-
-    Best-effort : un dépôt sans labs ``vm``, un provider sans backend snapshot
-    ou un hyperviseur muet ne doivent pas empêcher une destruction.
-    """
-    from .infra import snapshot as snapshot_infra
-
-    hosts = snapshot_infra.host_names(repo_meta)
-    if not hosts:
-        return
-    try:
-        retires = snapshot_infra.purge(repo_meta, hosts)
-    except Exception as exc:  # noqa: BLE001 — la destruction prime sur le ménage
-        warn(_("snapshot_purge_failed", error=str(exc)))
-        return
-    if retires:
-        info(_("snapshot_purge_done", count=len(retires)))
-        for chemin in retires:
-            info(f"  − {chemin}")
-
-
-def _handle_orphans_after_destroy(
-    repo_meta: RepoMetadata, *, assume_yes: bool
-) -> bool:
-    """Retire les machines que Terraform a laissées derrière lui, ou le dit.
-
-    Le retrait est **confirmé**, jamais implicite : rien ne garantit à dsoxlab
-    qu'un domaine homonyme d'un ``infra.hosts[].name`` soit bien le sien, et
-    une machine dé-définie ne revient pas. ``--yes`` vaut confirmation — c'est
-    déjà lui qui a autorisé la destruction de cette infrastructure, et ces
-    machines-là en font partie par leur nom.
-
-    Returns:
-        ``True`` si l'hyperviseur ne porte plus rien de ce dépôt et que la
-        commande peut sortir en succès. ``False`` s'il reste des machines :
-        l'appelant doit alors sortir en code non nul, faute de quoi il
-        annoncerait une destruction qui n'a pas eu lieu.
-    """
-    from .infra import terraform as tf
-
-    scan = tf.find_orphan_domains(repo_meta)
-    if scan.reason:
-        warn(_("orphan_check_skipped", error=scan.reason))
-    if not scan.orphans:
-        return True
-
-    noms = sorted(scan.orphans)
-    warn(_("destroy_orphan_domains", hosts=", ".join(noms)))
-    if not assume_yes and not typer.confirm(_("confirm_destroy_orphans")):
-        info(_("destroy_orphan_kept", cmd=_undefine_command(scan.orphans)))
-        return False
-
-    restants: dict[str, str] = {}
-    for fqdn in noms:
-        domaine = scan.orphans[fqdn]
-        try:
-            libvirt.remove_domain(domaine)
-        except CommandError as exc:
-            restants[fqdn] = domaine
-            error(_("destroy_orphan_failed", host=domaine,
-                    error=exc.result.stderr.strip() or str(exc)))
-    if restants:
-        info(_("destroy_orphan_kept", cmd=_undefine_command(restants)))
-        return False
-    success(_("destroy_orphan_removed", hosts=", ".join(noms)))
-    return True
-
-
-def _run_ansible_with_progress(
-    playbook_path: Path, runner: Callable[[EventCallback], Any]
-) -> None:
-    """Exécute un playbook Ansible en streamant les events vers Rich.
-
-    ``runner`` est une lambda qui prend un callback ``on_event`` et
-    invoque ``run_lab(lab, on_event=cb)`` ou équivalent
-    (``clean_lab``/``reset_lab`` côté services).
-
-    Events ansible-runner consommés :
-    - ``playbook_on_task_start`` : MàJ description avec nom de la tâche
-    - ``runner_on_ok`` / ``runner_on_failed`` / ``runner_on_unreachable``
-      / ``runner_on_skipped`` : ✔/✘/⊘ par tâche-host
-    - ``playbook_on_stats`` : récap final (silencieux, le caller utilise
-      les stats côté PlaybookResult).
-    """
-    from rich.progress import (
-        BarColumn,
-        MofNCompleteColumn,
-        Progress,
-        SpinnerColumn,
-        TextColumn,
-        TimeElapsedColumn,
-    )
-
-    state: dict[str, Any] = {"done": 0, "current_task": "", "playbook": playbook_path.name}
-
-    with Progress(
-        SpinnerColumn(),
-        TextColumn("[bold]{task.description}"),
-        BarColumn(bar_width=30),
-        MofNCompleteColumn(),
-        TimeElapsedColumn(),
-        transient=False,
-        console=console,
-    ) as progress:
-        task = progress.add_task(f"Running {playbook_path.name}…", total=None)
-
-        def on_event(event: dict[str, Any]) -> None:
-            etype = event.get("event")
-            data = event.get("event_data", {}) or {}
-
-            if etype == EVENT_INTERRUPT:
-                progress.console.print(_(
-                    "interrupt_notice_first" if event.get("count", 1) == 1
-                    else "interrupt_notice_second"
-                ))
-            elif etype == "playbook_on_task_start":
-                task_name = data.get("name") or data.get("task", "")
-                state["current_task"] = task_name
-                progress.update(task, description=_("progress_ansible_task", task=task_name))
-            elif etype == "runner_on_ok":
-                host = data.get("host", "?")
-                task_name = data.get("task", state["current_task"])
-                changed = data.get("res", {}).get("changed", False)
-                state["done"] += 1
-                progress.update(task, completed=state["done"])
-                marker = "[yellow]●[/yellow]" if changed else "[green]✔[/green]"
-                tag = "changed" if changed else "ok"
-                progress.console.print(
-                    f"  {marker} {host}: {task_name} [dim]({tag})[/dim]"
-                )
-            elif etype == "runner_on_skipped":
-                host = data.get("host", "?")
-                task_name = data.get("task", state["current_task"])
-                state["done"] += 1
-                progress.update(task, completed=state["done"])
-                progress.console.print(
-                    f"  [dim]⊘[/dim] {host}: {task_name} [dim](skipped)[/dim]"
-                )
-            elif etype == "runner_on_failed":
-                host = data.get("host", "?")
-                task_name = data.get("task", state["current_task"])
-                msg = data.get("res", {}).get("msg", "")
-                progress.console.print(
-                    f"  [red]✘ {host}: {task_name}[/red]"
-                )
-                if msg:
-                    progress.console.print(f"    [red]{msg}[/red]")
-            elif etype == "runner_on_unreachable":
-                host = data.get("host", "?")
-                msg = data.get("res", {}).get("msg", "host unreachable")
-                progress.console.print(
-                    f"  [red]✘ {host}: UNREACHABLE — {msg}[/red]"
-                )
-
-        runner(on_event)
-        progress.update(
-            task,
-            description=_("progress_playbook_done", playbook=playbook_path.name),
-        )
-
-
-def _run_terraform_init_with_spinner(runner: Callable[[EventCallback], Any]) -> None:
-    """Lance terraform init avec un spinner (téléchargement provider).
-
-    ``runner`` reçoit un callback ``on_event`` et invoque
-    ``tf.init(repo_meta, on_event=cb)``.
-    """
-    from rich.progress import Progress, SpinnerColumn, TextColumn, TimeElapsedColumn
-
-    with Progress(
-        SpinnerColumn(),
-        TextColumn("[bold]{task.description}"),
-        TimeElapsedColumn(),
-        transient=True,
-        console=console,
-    ) as progress:
-        task = progress.add_task("Terraform init…", total=None)
-
-        def on_event(event: dict[str, Any]) -> None:
-            etype = event.get("type")
-            # Quelques events du mode init -json
-            if etype == "log" or etype == "diagnostic":
-                level = event.get("@level", "info")
-                msg = event.get("@message", "")
-                if level == "error":
-                    progress.console.print(f"  [red]{msg}[/red]")
-                elif "Installing" in msg or "Finding" in msg or "Reusing" in msg:
-                    progress.update(task, description=msg.strip())
-
-        runner(on_event)
-        progress.update(task, description=_("progress_tf_init_done"))
-
-
-def _run_terraform_with_progress(
-    operation: str, runner: Callable[[EventCallback], Any]
-) -> Any:
-    """Exécute un terraform apply/destroy en streamant via Rich Progress.
-
-    ``runner`` est une lambda qui prend un callback ``on_event`` et
-    invoque ``tf.apply`` ou ``tf.destroy``. Retourne ce que ``runner``
-    retourne (ProvisionResult pour apply, None pour destroy).
-
-    La progress bar :
-    - liste les ressources à créer/détruire (via planned_change)
-    - avance à chaque apply_complete
-    - imprime ✔/✘ par ressource avec durée
-    - capture les diagnostics d'erreur pour le message final
-    """
-    from rich.progress import (
-        BarColumn,
-        MofNCompleteColumn,
-        Progress,
-        SpinnerColumn,
-        TextColumn,
-        TimeElapsedColumn,
-    )
-
-    # ``in_flight`` : ressources avec apply_start mais pas encore
-    # apply_complete (utile pour afficher ce qui prend du temps)
-    state: dict[str, Any] = {
-        "total": 0,
-        "done": 0,
-        "result": None,
-        "in_flight": {},  # addr → elapsed_seconds (depuis apply_progress)
-    }
-
-    label_action = "Creating" if operation == "provision" else "Destroying"
-    counted_action = "create" if operation == "provision" else "delete"
-
-    def _short_addr(addr: str) -> str:
-        """outscale_security_group_rule.bastion_ssh_in → bastion_ssh_in."""
-        return addr.split(".", 1)[-1] if "." in addr else addr
-
-    def _refresh_description() -> str:
-        """Description = liste des in-flight (avec elapsed s'il y en a)."""
-        if not state["in_flight"]:
-            return f"{label_action}…"
-        items = sorted(state["in_flight"].items(), key=lambda kv: -kv[1])
-        # Affiche jusqu'à 2 ressources pour ne pas surcharger
-        head = items[:2]
-        rest = len(items) - len(head)
-        parts = [f"{_short_addr(a)} ({s}s)" for a, s in head]
-        suffix = f" +{rest}" if rest > 0 else ""
-        return f"{label_action} {', '.join(parts)}{suffix}"
-
-    with Progress(
-        SpinnerColumn(),
-        TextColumn("[bold]{task.description}"),
-        BarColumn(bar_width=30),
-        MofNCompleteColumn(),
-        TimeElapsedColumn(),
-        transient=False,
-        console=console,
-    ) as progress:
-        task = progress.add_task("Planning…", total=None)
-
-        def on_event(event: dict[str, Any]) -> None:
-            etype = event.get("type")
-            if etype == "planned_change":
-                action = event.get("change", {}).get("action")
-                if action == counted_action:
-                    state["total"] += 1
-                    progress.update(task, total=state["total"])
-            elif etype == "apply_start":
-                hook = event.get("hook", {})
-                addr = hook.get("resource", {}).get("addr", "?")
-                state["in_flight"][addr] = 0
-                progress.update(task, description=_refresh_description())
-            elif etype == "apply_progress":
-                # Émis périodiquement par Terraform pour les ressources
-                # qui prennent du temps (elapsed_seconds croissant).
-                hook = event.get("hook", {})
-                addr = hook.get("resource", {}).get("addr", "?")
-                elapsed = int(hook.get("elapsed_seconds", 0))
-                if addr in state["in_flight"]:
-                    state["in_flight"][addr] = elapsed
-                    progress.update(task, description=_refresh_description())
-            elif etype == "apply_complete":
-                hook = event.get("hook", {})
-                addr = hook.get("resource", {}).get("addr", "?")
-                elapsed = hook.get("elapsed_seconds", 0)
-                state["in_flight"].pop(addr, None)
-                state["done"] += 1
-                progress.update(
-                    task,
-                    completed=state["done"],
-                    description=_refresh_description(),
-                )
-                progress.console.print(
-                    f"  [green]✔[/green] {_short_addr(addr)} [dim]({elapsed:.0f}s)[/dim]"
-                )
-            elif etype == "apply_errored":
-                hook = event.get("hook", {})
-                addr = hook.get("resource", {}).get("addr", "?")
-                state["in_flight"].pop(addr, None)
-                progress.update(task, description=_refresh_description())
-                progress.console.print(f"  [red]✘ {_short_addr(addr)}[/red]")
-            elif etype == "diagnostic":
-                if event.get("@level") == "error":
-                    diag = event.get("diagnostic", {})
-                    summary = diag.get("summary", "")
-                    progress.console.print(f"  [red]Error:[/red] {summary}")
-            elif etype == EVENT_INTERRUPT:
-                progress.console.print(_(
-                    "interrupt_notice_first" if event.get("count", 1) == 1
-                    else "interrupt_notice_second"
-                ))
-
-        state["result"] = runner(on_event)
-        # Une fois terminé, fixe la barre à 100 % avec un label final
-        if state["total"] > 0:
-            progress.update(
-                task,
-                description=_("progress_action_done", action=label_action),
-                completed=state["total"],
-            )
-        else:
-            progress.update(
-                task,
-                description=_("progress_nothing_to_do"),
-                total=1,
-                completed=1,
-            )
-
-    return state["result"]
-
-
-@app.command("status", help=_("cmd_lab_status_help"))
-def status(
-    lab_id: Annotated[str | None, typer.Argument(help=_("arg_lab_id_optionnel"))] = None,
-    lab_home: LabHomeOption = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    """Où en est le lab actif, ou celui qu'on nomme."""
-    from .services.lab_state import calculer
-
-    root = _root(lab_home)
-    lang = _lang(root)
-    ctx = read_context(root)
-    effective_id = lab_id or ctx.active_lab
-
-    if not effective_id:
-        # Sans lab actif, dire quoi faire — et nommer la commande qui portait
-        # ce nom jusqu'en 0.1.67, pour qui la cherche encore ici.
-        if as_json:
-            machine.emit({"schema": 1, "lab": None, "state": None})
-            return
-        info(_("status_aucun_lab_actif"))
-        info(_("status_voir_infra"))
-        return
-
-    try:
-        lab = _lab(root, effective_id, lang)
-    except ValueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    repo_meta = _read_repo(root)
-    repo_id = repo_meta.id if repo_meta else root.name
-    etat = calculer(root, lab, repo_id)
-
-    if as_json:
-        machine.emit({
-            "schema": 1,
-            "lab": lab.id,
-            "state": etat.state,
-            "label": etat.label,
-            "detail": etat.detail,
-            "best_score": etat.best_score,
-            "max_score": etat.max_score,
-        })
-        return
-
-    from .reporting.console import print_lab_state
-
-    print_lab_state(etat)
-
-
-@infra_app.command("status", help=_("cmd_status_help"))
-def infra_status(
-    lab_home: LabHomeOption = None,
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    """Vérifie la connectivité SSH des hosts du meta.yml."""
-    from .infra.inventory import build_inventory, read_terraform_outputs
-
-    root = _root(lab_home)
-    repo_meta = _read_repo(root)
-    if repo_meta is None:
-        error(_("provision_no_meta", root=root))
-        raise typer.Exit(1)
-
-    if not repo_meta.infra.hosts:
-        # Un dépôt sans infrastructure est un cas normal (catalogue 100 % shell),
-        # pas une erreur. En mode machine il faut tout de même un document :
-        # une phrase Rich et un code 0 laissaient l'appelant sans rien à lire.
-        if as_json:
-            machine.emit({
-                "provider": None,
-                "hosts": [],
-                "summary": {"reachable": 0, "total": 0},
-            })
-            return
-        info(_("status_no_hosts"))
-        return
-
-    provider = _require_provider(repo_meta)
-
-    from .infra.inventory import bastion_info
-
-    tf_outputs = read_terraform_outputs(repo_meta)
-    inventory = build_inventory(repo_meta, terraform_outputs=tf_outputs)
-    hosts_dict = inventory["all"]["children"]["labenv"]["hosts"]
-    bastion = bastion_info(tf_outputs, repo_meta=repo_meta)
-
-    ssh_key = repo_meta.path / "ssh" / "id_ed25519"
-    if not ssh_key.is_file():
-        error(_("status_no_key", path=ssh_key))
-        raise typer.Exit(1)
-
-    # On rafraîchit le fragment SSH au passage : « provision » ne peut pas
-    # toujours être rejoué (le provider libvirt refuse certaines mises à jour
-    # avec « Update Not Supported »), et l'apprenant se retrouverait alors sans
-    # moyen de régénérer sa configuration. « status » est la commande qu'on
-    # lance justement quand on doute de l'état de l'infra.
-    from .infra.inventory import write_user_ssh_config
-
-    try:
-        write_user_ssh_config(inventory, repo_meta)
-    except OSError as exc:
-        if not as_json:
-            warn(_("ssh_fragment_failed", error=str(exc)))
-
-    hotes: list[dict[str, Any]] = []
-    if not as_json:
-        info(_("status_checking", count=len(hosts_dict)))
-    if bastion:
-        info(_("status_via_bastion",
-               bastion=bastion["fqdn"] or bastion["public_ip"]))
-
-    # ── Interrogation de l'hyperviseur ───────────────────────────────────────
-    # Elle est PARESSEUSE : tant que tout répond, il n'y a rien à diagnostiquer
-    # et rien à demander. Elle n'a lieu qu'au premier hôte muet, et une seule
-    # fois pour toute la commande.
-    interrogeable = libvirt.supports_domain_state(provider)
-    domaines_connus: list[str] | None = None
-    hyperviseur_tente = False
-    hyperviseur_erreur: str | None = None
-
-    def _etat_domaine(fqdn: str) -> libvirt.DomainStatus | None:
-        """Ce que l'hyperviseur dit de cet hôte, ou ``None`` s'il ne dit rien.
-
-        ``None`` signifie « je n'ai pas pu regarder », jamais « rien n'existe » :
-        le diagnostic retombe alors sur ce que SSH sait, sans jamais conclure à
-        une machine absente.
-        """
-        nonlocal domaines_connus, hyperviseur_tente, hyperviseur_erreur
-        if not interrogeable:
-            return None
-        if not hyperviseur_tente:
-            hyperviseur_tente = True
-            try:
-                domaines_connus = libvirt.list_domains()
-            except CommandError as exc:
-                hyperviseur_erreur = exc.result.stderr.strip() or str(exc)
-        if domaines_connus is None:
-            return None
-        try:
-            return libvirt.inspect_host(fqdn, known=domaines_connus)
-        except CommandError as exc:
-            logger.debug("état libvirt indisponible pour %s : %s", fqdn, exc)
-            return None
-
-    ok_count = 0
-    for fqdn, host_vars in sorted(hosts_dict.items()):
-        ip = host_vars["ansible_host"]
-        cmd = [
-            "ssh",
-            # -F /dev/null : ignore la config SSH perso de l'apprenant
-            # (~/.ssh/config) qui peut contenir un ProxyJump appliqué
-            # par pattern d'IP (ex: "Host 10.*" → bastion). Sans ça,
-            # un fqdn lab en 10.x.x.x peut être routé vers un bastion
-            # tiers qui ne répond pas → "Connection to UNKNOWN port
-            # 65535 timed out".
-            "-F", "/dev/null",
-            "-o", "ConnectTimeout=4", "-o", "BatchMode=yes",
-            "-o", "StrictHostKeyChecking=no",
-            "-o", "UserKnownHostsFile=/dev/null",
-            "-i", str(ssh_key),
-        ]
-        # ProxyCommand explicite (pas ProxyJump) pour pouvoir
-        # injecter -i sur le hop bastion. ProxyJump natif d'OpenSSH
-        # ne propage pas le -i et utilise la clé par défaut de
-        # l'utilisateur (~/.ssh/id_ed25519), qui ne matche pas notre
-        # keypair lab → "Permission denied (publickey)".
-        if bastion:
-            proxy_target = bastion["fqdn"] or bastion["public_ip"]
-            cmd += [
-                "-o",
-                (
-                    f"ProxyCommand=ssh -F /dev/null -W %h:%p "
-                    f"-i {ssh_key} "
-                    f"-o StrictHostKeyChecking=no "
-                    f"-o UserKnownHostsFile=/dev/null "
-                    f"{bastion['user']}@{proxy_target}"
-                ),
-            ]
-        cmd += [f"{host_vars.get('ansible_user', 'ansible')}@{ip}", "true"]
-        # LC_ALL=C : la raison de l'échec vient de ``strerror``, que la libc
-        # traduit. Sans ce verrou, « No route to host » devient une phrase
-        # différente sur chaque poste, et le diagnostic ne reconnaîtrait plus
-        # rien de ce que ssh lui dit.
-        # check=False : l'échec de ce ssh EST la mesure. Un hôte injoignable
-        # est un résultat à rapporter, pas un incident qui arrête le tour des
-        # autres hôtes.
-        result = subprocess.run(
-            cmd, capture_output=True, timeout=15,
-            env={**os.environ, "LC_ALL": "C"}, check=False,
-        )
-        joignable = result.returncode == 0
-        raison = None
-        if not joignable:
-            stderr_tail = result.stderr.decode(errors="replace").strip().splitlines()[-1:]
-            raison = stderr_tail[0] if stderr_tail else "timeout"
-        etat = None if joignable else _etat_domaine(fqdn)
-        hotes.append({
-            "fqdn": fqdn,
-            "ip": ip,
-            "reachable": joignable,
-            "reason": raison,
-            "domain": etat.domain if etat else None,
-            "domain_state": etat.state if etat else None,
-            "cause": host_diagnosis.diagnose(
-                reachable=joignable, reason=raison or "", status=etat
-            ),
-        })
-        if as_json:
-            ok_count += 1 if joignable else 0
-            continue
-        if joignable:
-            # `success` et `error` posent déjà leur propre marqueur : en
-            # rajouter un dans la chaîne donnait « ✘   ✘ hote.lab », que
-            # l'utilisateur lit comme un défaut de rendu.
-            success(f"  {fqdn} ({ip})")
-            ok_count += 1
-        else:
-            error(f"  {fqdn} ({ip}) : {raison}")
-
-    if as_json:
-        machine.emit({
-            "provider": provider,
-            # Un consommateur doit pouvoir distinguer « aucun domaine » d'un
-            # hyperviseur qui n'a rien répondu : sans ce bloc, les deux se
-            # ressembleraient à des `domain_state: null`.
-            "hypervisor": {
-                "queryable": interrogeable and hyperviseur_erreur is None,
-                "error": hyperviseur_erreur,
-            },
-            "hosts": hotes,
-            "summary": {"reachable": ok_count, "total": len(hosts_dict)},
-        })
-        if ok_count != len(hosts_dict):
-            raise typer.Exit(1)
-        return
-    if ok_count == len(hosts_dict):
-        success(_("status_all_ok", count=ok_count))
-        return
-
-    error(_("status_partial", ok=ok_count, total=len(hosts_dict), provider=provider))
-    # Dire pourquoi le diagnostic est limité, plutôt que de laisser croire que
-    # l'outil a regardé la machine alors qu'il n'a pu regarder que le réseau.
-    if not interrogeable:
-        warn(_("status_provider_not_inspectable", provider=provider))
-    elif hyperviseur_erreur is not None:
-        warn(_("status_hypervisor_unavailable", error=hyperviseur_erreur))
-    for hote in hotes:
-        if hote["reachable"]:
-            continue
-        info(f"  {hote['fqdn']} — {_diagnostic_message(hote)}")
-    raise typer.Exit(1)
-
-
-@app.command("ssh", help=_("cmd_ssh_help"))
-def ssh_cmd(
-    host: Annotated[str, typer.Argument(help=_("cmd_ssh_arg"))],
-    lab_home: LabHomeOption = None,
-) -> None:
-    """Ouvre une session SSH interactive sur un host du meta.yml.
-
-    Si l'infrastructure expose un bastion (output Terraform ``bastion``),
-    SSH passe automatiquement par ProxyJump (règle non contournable
-    pour les providers cloud — REFACTORING-PLAN §11.8).
-    """
-    from .infra.inventory import bastion_info, build_inventory, read_terraform_outputs
-
-    root = _root(lab_home)
-    repo_meta = _read_repo(root)
-    if repo_meta is None:
-        error(_("provision_no_meta", root=root))
-        raise typer.Exit(1)
-
-    # Accepter le nom court (alma-rhcsa-1) ou le FQDN (alma-rhcsa-1.lab)
-    target_fqdn: str | None = None
-    for h in repo_meta.infra.hosts:
-        if h.name == host or h.name.split(".", 1)[0] == host:
-            target_fqdn = h.name
-            break
-    if target_fqdn is None:
-        error(_("ssh_unknown_host", host=host,
-                hosts=", ".join(h.name for h in repo_meta.infra.hosts)))
-        raise typer.Exit(1)
-
-    tf_outputs = read_terraform_outputs(repo_meta)
-    inventory = build_inventory(repo_meta, terraform_outputs=tf_outputs)
-    host_vars = inventory["all"]["children"]["labenv"]["hosts"][target_fqdn]
-    ip = host_vars["ansible_host"]
-    bastion = bastion_info(tf_outputs, repo_meta=repo_meta)
-
-    ssh_key = repo_meta.path / "ssh" / "id_ed25519"
-    cmd = [
-        "ssh",
-        # -F /dev/null : ignore la config SSH perso de l'apprenant.
-        "-F", "/dev/null",
-        "-o", "StrictHostKeyChecking=no",
-        "-o", "UserKnownHostsFile=/dev/null",
-        "-i", str(ssh_key),
-    ]
-    if bastion:
-        proxy_target = bastion["fqdn"] or bastion["public_ip"]
-        # ProxyCommand explicite avec -i pour le hop bastion
-        # (OpenSSH ProxyJump n'applique pas -i aux hops).
-        # -F /dev/null aussi côté bastion pour la même raison.
-        cmd += [
-            "-o",
-            (
-                f"ProxyCommand=ssh -F /dev/null -W %h:%p "
-                f"-i {ssh_key} "
-                f"-o StrictHostKeyChecking=no "
-                f"-o UserKnownHostsFile=/dev/null "
-                f"{bastion['user']}@{proxy_target}"
-            ),
-        ]
-        info(_("ssh_via_bastion", host=target_fqdn, ip=ip,
-                bastion=proxy_target))
-    else:
-        info(_("ssh_connecting", host=target_fqdn, ip=ip))
-    cmd.append(f"{host_vars.get('ansible_user', 'ansible')}@{ip}")
-    os.execvp("ssh", cmd)  # noqa: S606 — exec direct de ssh sans shell, argv construit en interne
-
-
-# ── instructor : commandes formateur ─────────────────────────────────────────
-
-
-@instructor_app.command("bootstrap", help=_("cmd_instructor_bootstrap_help"))
-def instructor_bootstrap(
-    lab_home: LabHomeOption = None,
-) -> None:
-    """Génère la clé SSH du lab si absente et vérifie les prérequis.
-
-    Crée ``<repo>/ssh/id_ed25519`` (+ .pub) sans passphrase. La clé
-    publique est ensuite injectée dans le tfvars par dsoxlab provision
-    pour être propagée aux VMs via cloud-init.
-
-    Vérifie la présence de ``terraform`` et ``ansible-runner``.
-    """
-    root = _root(lab_home)
-
-    # Une clé privée ne se crée pas n'importe où. Sans `meta.yml`, `root` n'est
-    # pas un dépôt de labs : `get_lab_home()` a simplement retenu le répertoire
-    # courant en dernier recours. Le cas est vécu : lancée depuis le dépôt de
-    # l'outil, la commande y a déposé une paire de clés hors de tout .gitignore.
-    # Le hook `detect-private-key` l'aurait arrêtée au commit, mais un hook se
-    # contourne (`--no-verify`) et ne protège que ce dépôt-ci. La clé n'avait
-    # de toute façon rien à faire là : on refuse plutôt que de deviner.
-    if not (root / "meta.yml").is_file():
-        error(_("bootstrap_not_a_lab_repo", root=root))
-        raise typer.Exit(1)
-
-    ssh_dir = root / "ssh"
-    private_key = ssh_dir / "id_ed25519"
-    public_key = ssh_dir / "id_ed25519.pub"
-
-    if private_key.is_file() and public_key.is_file():
-        info(_("bootstrap_key_exists", path=private_key))
-    else:
-        info(_("bootstrap_generating_key", path=private_key))
-        ssh_dir.mkdir(parents=True, exist_ok=True)
-        ssh_dir.chmod(0o700)
-        # check=False : l'échec est traduit en message + Exit(2) juste en
-        # dessous. Une CalledProcessError sortirait la trace Python de
-        # ssh-keygen au visage du formateur, sans lui dire quoi faire.
-        result = subprocess.run(
-            [
-                "ssh-keygen",
-                "-t", "ed25519",
-                "-N", "",  # pas de passphrase
-                "-C", "dsoxlab-lab",
-                "-f", str(private_key),
-            ],
-            capture_output=True,
-            text=True,
-            check=False,
-        )
-        if result.returncode != 0:
-            error(_("bootstrap_keygen_failed", stderr=result.stderr.strip()))
-            raise typer.Exit(2)
-        success(_("bootstrap_key_created", path=private_key))
-
-    # Vérifie les outils requis
-    from .infra import ansible as ansible_infra
-    from .infra import terraform as tf
-
-    manquant = False
-
-    if not tf.is_available():
-        error(_("bootstrap_no_terraform"))
-        manquant = True
-    else:
-        info(_("bootstrap_terraform_ok"))
-
-    if not ansible_infra.is_available():
-        error(_("bootstrap_no_ansible_runner"))
-        manquant = True
-    else:
-        info(_("bootstrap_ansible_runner_ok"))
-
-    # Sortir en 0 après avoir affiché une erreur bloquante trompe autant un
-    # apprenant qui vérifie son code de retour qu'un script d'installation :
-    # la clé SSH est bien créée, mais rien ne pourra la provisionner. Le code
-    # de retour doit dire la même chose que l'écran.
-    if manquant:
-        raise typer.Exit(1)
-
-
-# ── demo ──────────────────────────────────────────────────────────────────────
-
-@app.command("demo", help=_("cmd_demo_help"))
-def demo(
-    force: Annotated[bool, typer.Option("--force", help=_("opt_demo_force"))] = False,
-) -> None:
-    """Installe le catalogue de démonstration et dit quoi faire ensuite."""
-    from .services.demo import DemoExistante, installer
-
-    try:
-        installation = installer(force=force)
-    except DemoExistante as exc:
-        # Ne pas écraser : ce répertoire porte la progression et les réponses.
-        error(_("demo_deja_installee", path=str(exc)))
-        info(_("demo_deja_installee_suite", path=str(exc)))
-        raise typer.Exit(1) from None
-    except OSError as exc:
-        error(_("demo_echec", error=str(exc)))
-        raise typer.Exit(1) from None
-
-    success(_("demo_installee", path=str(installation.racine)))
-
-    # La marche à suivre est construite depuis le catalogue réellement
-    # installé : elle ne peut donc pas décrire un lab qui n'y serait plus.
-    premier = installation.labs[0] if installation.labs else ""
-    info(_("demo_suite", path=str(installation.racine), lab=premier))
-
-
-# ── catalog ───────────────────────────────────────────────────────────────────
-
-@catalog_app.command("list", help=_("cmd_catalog_list_help"))
-def catalog_list(
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-) -> None:
-    """Les catalogues connus, et ceux qui sont installés."""
-    from .i18n import get_lang
-    from .reporting.console import print_catalogues
-    from .services.catalog import installes, lire_manifeste
-
-    connus = lire_manifeste()
-    poses = installes()
-
-    if as_json:
-        machine.emit({
-            "schema": 1,
-            "known": [
-                {"id": c.id, "repository": c.depot,
-                 "description": c.description(get_lang())}
-                for c in connus
-            ],
-            "installed": [
-                {"id": p.id, "path": str(p.racine),
-                 "active": p.actif, "repository": p.depot}
-                for p in poses
-            ],
-        })
-        return
-
-    print_catalogues(connus, poses, get_lang())
-
-
-@catalog_app.command("add", help=_("cmd_catalog_add_help"))
-def catalog_add(
-    reference: Annotated[str, typer.Argument(help=_("arg_catalog_reference"))],
-    force: Annotated[
-        bool, typer.Option("--force", help=_("opt_catalog_force"))
-    ] = False,
-) -> None:
-    """Installe un catalogue et le rend actif."""
-    from .services.catalog import CatalogueError, ajouter, resoudre
-
-    try:
-        identifiant, url = resoudre(reference)
-        info(_("catalog_installation", name=identifiant, url=url))
-        pose = ajouter(reference, force=force)
-    except CatalogueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    success(_("catalog_installe", name=pose.id, path=str(pose.racine)))
-    info(_("catalog_installe_suite"))
-
-
-@catalog_app.command("update", help=_("cmd_catalog_update_help"))
-def catalog_update(
-    identifiant: Annotated[
-        str | None, typer.Argument(help=_("arg_catalog_id"))
-    ] = None,
-) -> None:
-    """Met à jour un catalogue, ou tous ceux qui sont installés."""
-    from .services.catalog import CatalogueError, installes, mettre_a_jour
-
-    cibles = [identifiant] if identifiant else [p.id for p in installes()]
-    if not cibles:
-        info(_("catalog_aucun_installe"))
-        return
-
-    echecs = 0
-    for cible in cibles:
-        try:
-            detail = mettre_a_jour(cible)
-        except CatalogueError as exc:
-            error(str(exc))
-            echecs += 1
-            continue
-        # git dit « Already up to date. » quand il n'a rien fait : le répéter
-        # tel quel obligerait à lire de l'anglais dans une session française.
-        if "up to date" in detail.lower() or not detail:
-            info(_("catalog_a_jour", name=cible))
-        else:
-            success(_("catalog_mis_a_jour", name=cible, detail=detail))
-
-    # Une mise à jour ratée parmi plusieurs doit se voir dans le code de
-    # retour : un script qui enchaîne ne lit pas l'écran.
-    if echecs:
-        raise typer.Exit(1)
-
-
-@catalog_app.command("remove", help=_("cmd_catalog_remove_help"))
-def catalog_remove(
-    identifiant: Annotated[str, typer.Argument(help=_("arg_catalog_id"))],
-) -> None:
-    """Retire un catalogue installé."""
-    from .services.catalog import CatalogueError, retirer
-
-    try:
-        racine = retirer(identifiant)
-    except CatalogueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    success(_("catalog_retire", name=identifiant, path=str(racine)))
-
-
-@catalog_app.command("use", help=_("cmd_catalog_use_help"))
-def catalog_use(
-    identifiant: Annotated[str, typer.Argument(help=_("arg_catalog_id"))],
-) -> None:
-    """Choisit le catalogue actif."""
-    from .services.catalog import CatalogueError, definir_actif
-
-    try:
-        racine = definir_actif(identifiant)
-    except CatalogueError as exc:
-        error(str(exc))
-        raise typer.Exit(1) from None
-
-    success(_("catalog_actif_defini", name=identifiant, path=str(racine)))
-
-
-# ── support ───────────────────────────────────────────────────────────────────
-
-@app.command("support", help=_("cmd_support_help"))
-def support(
-    as_json: Annotated[bool, typer.Option("--json", help=_("opt_json"))] = False,
-    lignes: Annotated[
-        int, typer.Option("--log-lines", help=_("opt_support_log_lines"))
-    ] = 30,
-) -> None:
-    """Rapport de diagnostic anonymisé, prêt à coller dans une issue."""
-    from .services.support import collecter, en_markdown
-
-    rapport = collecter(lignes_journal=max(0, lignes))
-
-    if as_json:
-        machine.emit(rapport)
-        return
-
-    # `print` et non la console Rich : ce texte est fait pour être copié dans
-    # une issue. Rich l'habillerait de couleurs et le couperait à la largeur du
-    # terminal, ce qui casserait les tableaux Markdown une fois collés.
-    print(en_markdown(rapport))
-    info(_("support_hint"))
-
-
-# ── fullhelp ──────────────────────────────────────────────────────────────────
-
-@app.command("fullhelp", help=_("cmd_fullhelp_help"))
-def fullhelp() -> None:
-    print_fullhelp()
-
-
-# ── point d'entrée console ────────────────────────────────────────────────────
-
-def main() -> None:
-    """Point d'entrée de la commande ``dsoxlab``.
-
-    Enveloppe l'app Typer pour rendre les erreurs ATTENDUES en une phrase
-    actionnable, jamais en traceback Python. Une infrastructure non
-    provisionnée est une situation normale (premier lancement, après un
-    ``destroy``) : l'apprenant doit lire quoi faire, pas une pile d'appels.
-
-    Les erreurs inattendues, elles, continuent de remonter avec leur
-    traceback — c'est ce qu'on veut pour un vrai bug.
-    """
-    try:
-        app()
-    except Interrupted as exc:
-        # Interruption d'une commande qui n'avait pas de consigne de reprise à
-        # donner, ou survenue hors des étapes inventoriées. On dit au moins ce
-        # qui a été interrompu, et on sort en 130 plutôt qu'en 1.
-        warn(_(exc.message_key))
-        if exc.hard:
-            warn(_("interrupted_hard"))
-        raise SystemExit(EXIT_INTERRUPTED) from None
-    except InfraNotProvisioned:
-        error(_("infra_not_provisioned"))
-        raise SystemExit(1) from None
-    # Filet de dernier recours : les commandes qui lisent le catalogue passent
-    # par `_read_repo` ou `_catalogue`, qui disent déjà la même chose plus tôt.
-    # Celles qui liraient un meta.yml par un autre chemin ne doivent pas pour
-    # autant rendre un traceback là où une phrase suffit.
-    except UnsupportedSchemaVersion as exc:
-        error(_(
-            "schema_version_meta_too_new",
-            path=exc.source, found=exc.found, supported=exc.supported,
-        ))
-        raise SystemExit(1) from None
-    # Filet de dernier recours, comme au-dessus : une commande qui lirait un
-    # meta.yml par un autre chemin doit rendre la phrase, pas un traceback.
-    except ContractError as exc:
-        error(_phrase_contrat(exc))
-        raise SystemExit(1) from None
